Crate aster_kernel

Aster-nix is the Asterinas kernel, a safe, efficient unix-like operating system kernel built on top of OSTD and OSDK.

Modules

arch 🔒

context 🔒

The context that can be accessed from the current task, thread or process.

cpu 🔒

device 🔒

driver 🔒

error 🔒

events 🔒

fs 🔒

init 🔒

Kernel initialization.

ipc 🔒

net 🔒

prelude 🔒

process 🔒

sched 🔒

security 🔒

syscall 🔒

System call handlers.

thread 🔒

Posix thread implementation

time 🔒

util 🔒

vdso 🔒

Virtual Dynamic Shared Object (vDSO).

vm 🔒

Virtual memory (VM).

Macros

current

return current process

current_thread

Returns the current thread.

current_userspace

Gets the CurrentUserSpace from the current task.

impl_raw_sock_option_get_only

Impl RawSocketOption for a struct which is for only getsockopt and implements SocketOption.

impl_raw_sock_option_set_only

Impl RawSocketOption for a struct which is for only setsockopt and implements SocketOption.

impl_raw_socket_option

Impl RawSocketOption for a struct which implements SocketOption.

log_syscall_entry

return_errno

return_errno_with_message

Functions

__ostd_main 🔒

main 🔒