aster_kernel::context

Function check_vaddr_lowerbound

fn check_vaddr_lowerbound(va: Vaddr) -> Result<()>

Checks if the user space pointer is below the lowest userspace address.

If a pointer is below the lowest userspace address, it is likely to be a NULL pointer. Reading from or writing to a NULL pointer should trigger a segmentation fault.

If it is not checked here, a kernel page fault will happen and we would deny the access in the page fault handler anyway. It may save a page fault in some occasions. More importantly, double page faults may not be handled quite well on some platforms.