aster_kernel::device::evdev::file

Struct EvdevFile

pub(super) struct EvdevFile {
    inner: Arc<EvdevFileInner>,
    evdev: Weak<EvdevDevice>,
}

An opened file from an evdev device (EvdevDevice).

Fields
inner: Arc<EvdevFileInner>

Inner data (shared with the device).

evdev: Weak<EvdevDevice>

Weak reference to the evdev device that owns this evdev file.

Implementations

impl EvdevFile

pub(super) fn new( buffer_size: usize, evdev: Weak<EvdevDevice>, ) -> (Self, Producer<EvdevEvent, Arc<RingBuffer<EvdevEvent>>>)

pub(super) fn inner(&self) -> &Arc<EvdevFileInner>

fn process_events( &self, max_events: usize, writer: &mut VmWriter<'_>, ) -> Result<usize, Error>

Processes events and writes them to the writer.

Returns the total number of bytes written, or Errno::EAGAIN if no events are available.

fn check_io_events(&self) -> IoEvents

fn upgrade_device(&self) -> Result<Arc<EvdevDevice>, Error>

fn set_clock(&self, clock: EvdevClock) -> Result<(), Error>