aster_kernel::device::registry

Module char

A subsystem for character devices (or char devices for short).

Structs

MajorIdOwner

An owned major ID.

Constants

DYNAMIC_MAJOR_ID_RANGES 🔒

The ranges of free char majors.

MAX_MAJOR

The maximum value of the major device ID of a char device.

Statics

DEVICE_REGISTRY 🔒

MAJORS 🔒

Functions

acquire_major

Acquires a major ID.

allocate_major

Allocates a major ID.

collect_all

Collects all char devices.

init_in_first_process 🔒

lookup 🔒

Looks up a char device of a given device ID.

register

Registers a new char device.

unregister

Unregisters an existing char device, returning the device if found.