aster_kernel::device::tty

Module device

TTY devices.

This module implements TTY devices such as /dev/tty0, /dev/tty, and /dev/console.

Reference: https://www.kernel.org/doc/html/latest/admin-guide/devices.html

Structs

SystemConsole

Corresponds to /dev/console in the file system. This device represents a console to which system messages will be sent.

Tty0Device

Corresponds to /dev/tty0 in the file system. This device represents the active virtual terminal.

TtyDevice

Corresponds to /dev/tty in the file system. This device represents the controlling terminal of the session of the current process.

Functions

init_in_first_process 🔒