aster_kernel::fs::device

Trait Device

pub trait Device:
    Send
    + Sync
    + 'static {
    // Required methods
    fn type_(&self) -> DeviceType;
    fn id(&self) -> DeviceId;
    fn devtmpfs_path(&self) -> Option<String>;
    fn open(&self) -> Result<Box<dyn FileIo>, Error>;
}

The abstraction of a device.

Required Methods

fn type_(&self) -> DeviceType

Returns the device type.

fn id(&self) -> DeviceId

Returns the device ID.

fn devtmpfs_path(&self) -> Option<String>

Returns the path where the device should appear in devtmpfs (usually under /dev), if any.

fn open(&self) -> Result<Box<dyn FileIo>, Error>

Opens the device, returning a file-like object that the userspace can interact with by doing I/O.