aster_kernel::fs::exfat

Module dentry

Structs

ExfatBitmapDentry ğŸ”’

ExfatDeletedDentry ğŸ”’

ExfatDentryIterator ğŸ”’

ExfatDentrySet ğŸ”’

A set of dentries that collectively describe a file or folder. Root directory cannot be represented as an ordinal dentryset.

ExfatFileDentry ğŸ”’

On-disk dentry formats

ExfatGenericPrimaryDentry ğŸ”’

ExfatGenericSecondaryDentry ğŸ”’

ExfatName ğŸ”’

ExfatNameDentry ğŸ”’

ExfatStreamDentry ğŸ”’

ExfatUpcaseDentry ğŸ”’

ExfatVendorAllocDentry ğŸ”’

ExfatVendorExtDentry ğŸ”’

RawExfatDentry ğŸ”’

Enums

ExfatDentry ğŸ”’

ExfatValidateDentryMode ğŸ”’

Constants

DENTRY_SIZE ğŸ”’

EXFAT_ACL ğŸ”’

EXFAT_ACLTAB ğŸ”’

EXFAT_BITMAP ğŸ”’

EXFAT_FILE ğŸ”’

EXFAT_GUID ğŸ”’

EXFAT_INVAL ğŸ”’

EXFAT_NAME ğŸ”’

EXFAT_PADDING ğŸ”’

EXFAT_STREAM ğŸ”’

EXFAT_UNUSED ğŸ”’

EXFAT_UPCASE ğŸ”’

EXFAT_VENDOR_ALLOC ğŸ”’

EXFAT_VENDOR_EXT ğŸ”’

EXFAT_VOLUME ğŸ”’

Traits

Checksum

Type Aliases

UTF16Char