aster_kernel::fs::exfat::inode

Struct ExfatInodeInner

struct ExfatInodeInner {

    ino: u64,
    dentry_set_position: (ExfatChain, usize),
    dentry_set_size: usize,
    dentry_entry: u32,
    inode_type: InodeType,
    attr: FatAttr,
    start_chain: ExfatChain,
    size: usize,
    size_allocated: usize,
    atime: DosTimestamp,
    mtime: DosTimestamp,
    ctime: DosTimestamp,
    num_sub_inodes: u32,
    num_sub_dirs: u32,
    name: ExfatName,
    is_deleted: bool,
    parent_hash: usize,
    fs: Weak<ExfatFs>,
    page_cache: PageCache,

}

Fields
ino: u64

Inode number.

dentry_set_position: (ExfatChain, usize)

Dentry set position in its parent directory.

dentry_set_size: usize

Dentry set size in bytes.

dentry_entry: u32

The entry number of the dentry.

inode_type: InodeType

Inode type, File or Dir.

attr: FatAttrstart_chain: ExfatChain

Start position on disk, this is undefined if the allocated size is 0.

size: usize

Valid size of the file.

size_allocated: usize

Allocated size, for directory, size is always equal to size_allocated.

atime: DosTimestamp

Access time, updated after reading.

mtime: DosTimestamp

Modification time, updated only on write.

ctime: DosTimestamp

Creation time.

num_sub_inodes: u32

Number of sub inodes.

num_sub_dirs: u32

Number of sub inodes that are directories.

name: ExfatName

ExFAT uses UTF-16 encoding, rust use utf-8 for string processing.

is_deleted: bool

Flag for whether the inode is deleted.

parent_hash: usize

The hash of its parent inode.

fs: Weak<ExfatFs>

A pointer to exFAT fs.

page_cache: PageCache

Important: To enlarge the page_cache, we need to update the page_cache size before we update the size of inode, to avoid extra data read. To shrink the page_cache, we need to update the page_cache size after we update the size of inode, to avoid extra data write.

Implementations

impl ExfatInodeInner

fn hash_index(&self) -> usize

The hash_value to index inode. This should be unique in the whole fs. Currently use dentry set physical position as hash value except for root(0).

fn get_parent_inode(&self) -> Option<Arc<ExfatInode>>

fn get_sector_id(&self, sector_id: usize) -> Result<usize, Error>

Get physical sector id from logical sector id for this Inode.

fn get_physical_cluster(&self, logical: u32) -> Result<u32, Error>

Get the physical cluster id from the logical cluster id in the inode.

fn num_clusters(&self) -> u32

The number of clusters allocated.

fn is_sync(&self) -> bool

fn fs(&self) -> Arc<ExfatFs>

fn is_empty_dir(&self) -> Result<bool, Error>

Only valid for directory, check if the dir is empty.

fn make_mode(&self) -> InodeMode

fn count_num_sub_inode_and_dir( &self, fs_guard: &MutexGuard<'_, ()>, ) -> Result<(usize, usize), Error>

fn resize( &mut self, new_size: usize, fs_guard: &MutexGuard<'_, ()>, ) -> Result<(), Error>

Resize current inode to new_size. The size_allocated field in inode can be enlarged, while the size field will not.

fn write_inode( &self, sync: bool, fs_guard: &MutexGuard<'_, ()>, ) -> Result<(), Error>

Update inode information back to the disk to sync this inode. Should lock the file system before calling this function.

fn visit_sub_inodes( &self, offset: usize, dir_cnt: usize, visitor: &mut dyn DirentVisitor, fs_guard: &MutexGuard<'_, ()>, ) -> Result<(usize, usize), Error>

Read all sub-inodes from the given position(offset) in this directory. The number of inodes to read is given by dir_cnt. Return (the new offset after read, the number of sub-inodes read).

fn visit_sub_inode( &self, file_dentry: &ExfatFileDentry, iter: &mut ExfatDentryIterator<'_>, offset: usize, visitor: &mut dyn DirentVisitor, fs_guard: &MutexGuard<'_, ()>, ) -> Result<usize, Error>

Visit a sub-inode at offset. Return the dentry-set size of the sub-inode. Dirent visitor will extract information from the inode.

fn lookup_by_name( &self, target_name: &str, case_sensitive: bool, fs_guard: &MutexGuard<'_, ()>, ) -> Result<Arc<ExfatInode>, Error>

Look up a target with “name”, cur inode represent a dir. Return (target inode, dentries start offset, dentry set size). No inode should hold the write lock.

fn delete_associated_secondary_clusters( &mut self, dentry: &ExfatDentry, fs_guard: &MutexGuard<'_, ()>, ) -> Result<(), Error>

fn free_all_clusters( &mut self, fs_guard: &MutexGuard<'_, ()>, ) -> Result<(), Error>

fn sync_metadata(&self, fs_guard: &MutexGuard<'_, ()>) -> Result<(), Error>

fn sync_data(&self, fs_guard: &MutexGuard<'_, ()>) -> Result<(), Error>

fn sync_all(&self, fs_guard: &MutexGuard<'_, ()>) -> Result<(), Error>

fn update_metadata_for_delete(&mut self, is_dir: bool)

Update the metadata for current directory after a delete. Set is_dir if the deleted file is a directory.

fn update_atime(&mut self) -> Result<(), Error>

fn update_atime_and_mtime(&mut self) -> Result<(), Error>