aster_kernel::fs::ext2

Module prelude

Structs

BioSegment 🔒

BioSegment is the basic memory unit of a block I/O request.

BioWaiter 🔒

A waiter for Bio submissions.

Duration 🔒

A Duration type to represent a span of time, typically used for system timeouts.

Frame 🔒

A smart pointer to a frame.

FrameAllocOptions 🔒

Options for allocating physical memory frames.

Range 🔒

A (half-open) range bounded inclusively below and exclusively above (start..end).

RwMutex 🔒

A mutex that provides data access to either one writer or many readers.

RwMutexReadGuard 🔒

A guard that provides immutable data access.

RwMutexWriteGuard 🔒

A guard that provides mutable data access.

Segment 🔒

A contiguous range of homogeneous physical memory frames.

Enums

BioDirection 🔒

The direction of a bio request.

BioStatus 🔒

The status of Bio.

Constants

BLOCK_SIZE 🔒

SECTOR_SIZE 🔒

Traits

AlignExt 🔒

An extension trait for Rust integer types, including u8, u16, u32, u64, and usize, to provide methods to make integers aligned to a power of two.

BlockDevice 🔒

Deref 🔒

Used for immutable dereferencing operations, like *v.

DerefMut 🔒

Used for mutable dereferencing operations, like in *v = 1;.

VmIo 🔒

A trait that enables reading/writing data from/to a VM object, e.g., USegment, Vec<UFrame> and UFrame.

Type Aliases

Bid 🔒

The block index used in the filesystem.

USegment 🔒

A contiguous range of homogeneous untyped physical memory frames that have any metadata.