aster_kernel::fs::ext2::prelude

Struct BioSegment

pub(in fs::ext2) struct BioSegment {
    inner: Arc<BioSegmentInner>,
}

BioSegment is the basic memory unit of a block I/O request.

Fields
inner: Arc<BioSegmentInner>
Implementations

impl BioSegment

pub fn alloc(nblocks: usize, direction: BioDirection) -> BioSegment

Allocates a new BioSegment with the wanted blocks count and the bio direction.

pub fn new_from_segment( segment: Segment<dyn AnyUFrameMeta>, direction: BioDirection, ) -> BioSegment

Constructs a new BioSegment with a given USegment and the bio direction.

pub fn nbytes(&self) -> usize

Returns the number of bytes.

pub fn nsectors(&self) -> BlockId<aster_block::::id::Sid::{constant#0}>

Returns the number of sectors.

pub fn nblocks(&self) -> usize

Returns the number of blocks.

pub fn offset_within_first_block(&self) -> usize

Returns the offset (in bytes) within the first block.

pub fn inner_dma_slice(&self) -> &Slice<Arc<DmaStream>>

Returns the inner DMA slice.