aster_kernel::fs::ext2::prelude

Struct Frame

#[repr(transparent)]
pub(in fs::ext2) struct Frame<M>
where
    M: AnyFrameMeta + ?Sized,
{
    ptr: *const MetaSlot,
    _marker: PhantomData<M>,
}

A smart pointer to a frame.

A frame is a contiguous range of bytes in physical memory. The Frame type is a smart pointer to a frame that is reference-counted.

Frames are associated with metadata. The type of the metadata M is determines the kind of the frame. If M implements AnyUFrameMeta, the frame is a untyped frame. Otherwise, it is a typed frame.

Fields
ptr: *const MetaSlot_marker: PhantomData<M>
Implementations

impl<M> Frame<M>
where M: AnyFrameMeta,

pub fn from_unused(paddr: usize, metadata: M) -> Result<Frame<M>, GetFrameError>

Gets a Frame with a specific usage from a raw, unused page.

The caller should provide the initial metadata of the page.

If the provided frame is not truly unused at the moment, it will return an error. If wanting to acquire a frame that is already in use, use Frame::from_in_use instead.

pub fn meta(&self) -> &M

Gets the metadata of this page.

impl Frame<dyn AnyFrameMeta>

pub fn from_in_use( paddr: usize, ) -> Result<Frame<dyn AnyFrameMeta>, GetFrameError>

Gets a dynamically typed Frame from a raw, in-use page.

If the provided frame is not in use at the moment, it will return an error.

The returned frame will have an extra reference count to the frame.

impl<M> Frame<M>
where M: AnyFrameMeta + ?Sized,

pub const fn map_level(&self) -> u8

Gets the map level of this page.

This is the level of the page table entry that maps the frame, which determines the size of the frame.

Currently, the level is always 1, which means the frame is a regular page frame.

pub fn dyn_meta(&self) -> &(dyn AnyFrameMeta + 'static)

Gets the dyncamically-typed metadata of this frame.

If the type is known at compile time, use Frame::meta instead.

pub fn reference_count(&self) -> u64

Gets the reference count of the frame.

It returns the number of all references to the frame, including all the existing frame handles (Frame, Frame<dyn AnyFrameMeta>), and all the mappings in the page table that points to the frame.

Safety

The function is safe to call, but using it requires extra care. The reference count can be changed by other threads at any time including potentially between calling this method and acting on the result.

pub fn borrow(&self) -> FrameRef<'_, M>

Borrows a reference from the given frame.

impl Frame<dyn AnyFrameMeta>

pub fn from_unsized<M>(frame: Frame<M>) -> Frame<dyn AnyFrameMeta>
where M: AnyFrameMeta + ?Sized,

Converts a Frame with a specific metadata type into a Frame<dyn AnyFrameMeta>.

This exists because:

impl<M: AnyFrameMeta + ?Sized> From<Frame<M>> for Frame<dyn AnyFrameMeta>

will conflict with impl<T> core::convert::From<T> for T in crate core.