aster_kernel::fs::ext2::prelude

Struct Segment

#[repr(transparent)]
pub(in fs::ext2) struct Segment<M>
where
    M: AnyFrameMeta + ?Sized,
{
    range: Range<usize>,
    _marker: PhantomData<M>,
}

A contiguous range of homogeneous physical memory frames.

This is a handle to multiple contiguous frames. It will be more lightweight than owning an array of frame handles.

The ownership is achieved by the reference counting mechanism of frames. When constructing a Segment, the frame handles are created then forgotten, leaving the reference count. When dropping a it, the frame handles are restored and dropped, decrementing the reference count.

All the metadata of the frames are homogeneous, i.e., they are of the same type.

Fields
range: Range<usize>_marker: PhantomData<M>
Implementations

impl<M> Segment<M>
where M: AnyFrameMeta,

pub fn from_unused<F>( range: Range<usize>, metadata_fn: F, ) -> Result<Segment<M>, GetFrameError>
where F: FnMut(usize) -> M,

Creates a new Segment from unused frames.

The caller must provide a closure to initialize metadata for all the frames. The closure receives the physical address of the frame and returns the metadata, which is similar to [core::array::from_fn].

It returns an error if:

the physical address is invalid or not aligned;

any of the frames cannot be created with a specific reason.

Panics

It panics if the range is empty.

impl<M> Segment<M>
where M: AnyFrameMeta + ?Sized,

pub fn slice(&self, range: &Range<usize>) -> Segment<M>

Gets an extra handle to the frames in the byte offset range.

The sliced byte offset range in indexed by the offset from the start of the contiguous frames. The resulting frames holds extra reference counts.

Panics

The function panics if the byte offset range is out of bounds, or if any of the ends of the byte offset range is not base-page aligned.

impl Segment<dyn AnyFrameMeta>

pub fn from_unsized<M>(segment: Segment<M>) -> Segment<dyn AnyFrameMeta>
where M: AnyFrameMeta + ?Sized,

Converts a Segment with a specific metadata type into a Segment<dyn AnyFrameMeta>.

This exists because:

impl<M: AnyFrameMeta + ?Sized> From<Segment<M>> for Segment<dyn AnyFrameMeta>

will conflict with impl<T> core::convert::From<T> for T in crate core.

See also Frame::from_unsized.