aster_kernel::fs::ext2::prelude

Trait BlockDevice

pub(in fs::ext2) trait BlockDevice:
    Send
    + Sync
    + Any
    + Debug {
    // Required methods
    fn enqueue(&self, bio: SubmittedBio) -> Result<(), BioEnqueueError>;
    fn metadata(&self) -> BlockDeviceMeta;
    fn name(&self) -> &str;
    fn id(&self) -> DeviceId;

    // Provided methods
    fn is_partition(&self) -> bool { ... }
    fn set_partitions(&self, _infos: Vec<Option<PartitionInfo>>) { ... }
    fn partitions(&self) -> Option<Vec<Arc<dyn BlockDevice>>> { ... }
}

Required Methods

fn enqueue(&self, bio: SubmittedBio) -> Result<(), BioEnqueueError>

Enqueues a new SubmittedBio to the block device.

fn metadata(&self) -> BlockDeviceMeta

Returns the metadata of the block device.

fn name(&self) -> &str

Returns the name of the block device.

fn id(&self) -> DeviceId

Returns the device ID of the block device.

Provided Methods

fn is_partition(&self) -> bool

Returns whether the block device is a partition.

fn set_partitions(&self, _infos: Vec<Option<PartitionInfo>>)

Sets the partitions of the block device.

fn partitions(&self) -> Option<Vec<Arc<dyn BlockDevice>>>

Returns the partitions of the block device.

Implementations

impl dyn BlockDevice

Implements several commonly used APIs for the block device to conveniently read and write block(s).

pub fn read_blocks( &self, bid: BlockId<aster_block::::id::Bid::{constant#0}>, bio_segment: BioSegment, ) -> Result<BioStatus, BioEnqueueError>

Synchronously reads contiguous blocks starting from the bid.

pub fn read_blocks_async( &self, bid: BlockId<aster_block::::id::Bid::{constant#0}>, bio_segment: BioSegment, ) -> Result<BioWaiter, BioEnqueueError>

Asynchronously reads contiguous blocks starting from the bid.

pub fn write_blocks( &self, bid: BlockId<aster_block::::id::Bid::{constant#0}>, bio_segment: BioSegment, ) -> Result<BioStatus, BioEnqueueError>

Synchronously writes contiguous blocks starting from the bid.

pub fn write_blocks_async( &self, bid: BlockId<aster_block::::id::Bid::{constant#0}>, bio_segment: BioSegment, ) -> Result<BioWaiter, BioEnqueueError>

Asynchronously writes contiguous blocks starting from the bid.

pub fn sync(&self) -> Result<BioStatus, BioEnqueueError>

Issues a sync request

impl dyn BlockDevice

pub fn write_bytes_async( &self, offset: usize, buf: &[u8], ) -> Result<BioWaiter, Error>

Asynchronously writes consecutive bytes of several sectors in size.

impl dyn BlockDevice

pub fn downcast_ref<T>(&self) -> Option<&T>
where T: BlockDevice,