aster_kernel::fs::file_handle

Trait FileLike

pub trait FileLike:
    Pollable
    + Send
    + Sync
    + Any {

    // Required methods
    fn path(&self) -> &Path;
    fn dump_proc_fdinfo(self: Arc<Self>, fd_flags: FdFlags) -> Box<dyn Display>;

    // Provided methods
    fn read(&self, writer: &mut VmWriter<'_>) -> Result<usize, Error> { ... }
    fn write(&self, reader: &mut VmReader<'_>) -> Result<usize, Error> { ... }
    fn read_at(
        &self,
        offset: usize,
        writer: &mut VmWriter<'_>,
    ) -> Result<usize, Error> { ... }
    fn write_at(
        &self,
        offset: usize,
        reader: &mut VmReader<'_>,
    ) -> Result<usize, Error> { ... }
    fn ioctl(&self, raw_ioctl: RawIoctl) -> Result<i32, Error> { ... }
    fn mappable(&self) -> Result<Mappable, Error> { ... }
    fn resize(&self, new_size: usize) -> Result<(), Error> { ... }
    fn status_flags(&self) -> StatusFlags { ... }
    fn set_status_flags(&self, _new_flags: StatusFlags) -> Result<(), Error> { ... }
    fn access_mode(&self) -> AccessMode { ... }
    fn seek(&self, seek_from: SeekFrom) -> Result<usize, Error> { ... }
    fn fallocate(
        &self,
        mode: FallocMode,
        offset: usize,
        len: usize,
    ) -> Result<(), Error> { ... }
    fn as_socket(&self) -> Option<&dyn Socket> { ... }

}

The basic operations defined on a file

Required Methods

fn path(&self) -> &Path

fn dump_proc_fdinfo(self: Arc<Self>, fd_flags: FdFlags) -> Box<dyn Display>

Dumps information to appear in the fdinfo file under procfs.

This method must not break atomic mode because it will be called with the file table’s spin lock held. There are two strategies for implementing this method:

If the necessary information can be obtained without breaking atomic mode, the method can collect and return the information directly. Arc<Self> should be dropped and should not appear in the returned Box<dyn Display>.

Otherwise, if the file can be dropped asynchronously in another process, the method can return a Box<dyn Display> containing the Arc<Self>, so that the information can be collected later in its Display::display() method, after dropping the file table’s spin lock.

Provided Methods

fn read(&self, writer: &mut VmWriter<'_>) -> Result<usize, Error>

fn write(&self, reader: &mut VmReader<'_>) -> Result<usize, Error>

fn read_at( &self, offset: usize, writer: &mut VmWriter<'_>, ) -> Result<usize, Error>

Read at the given file offset.

The file must be seekable to support read_at. Unlike read, read_at will not change the file offset.

fn write_at( &self, offset: usize, reader: &mut VmReader<'_>, ) -> Result<usize, Error>

Write at the given file offset.

The file must be seekable to support write_at. Unlike write, write_at will not change the file offset. If the file is append-only, the offset will be ignored.

fn ioctl(&self, raw_ioctl: RawIoctl) -> Result<i32, Error>

fn mappable(&self) -> Result<Mappable, Error>

Obtains the mappable object to map this file into the user address space.

If this file has a corresponding mappable object of Mappable, then it can be either an inode or an MMIO region.

fn resize(&self, new_size: usize) -> Result<(), Error>

fn status_flags(&self) -> StatusFlags

fn set_status_flags(&self, _new_flags: StatusFlags) -> Result<(), Error>

fn access_mode(&self) -> AccessMode

fn seek(&self, seek_from: SeekFrom) -> Result<usize, Error>

fn fallocate( &self, mode: FallocMode, offset: usize, len: usize, ) -> Result<(), Error>

fn as_socket(&self) -> Option<&dyn Socket>

Implementations

impl dyn FileLike

pub fn downcast_ref<T: FileLike>(&self) -> Option<&T>

pub fn read_bytes(&self, buf: &mut [u8]) -> Result<usize, Error>

pub fn write_bytes(&self, buf: &[u8]) -> Result<usize, Error>

pub fn read_bytes_at( &self, offset: usize, buf: &mut [u8], ) -> Result<usize, Error>

pub fn write_bytes_at(&self, offset: usize, buf: &[u8]) -> Result<usize, Error>

pub fn as_socket_or_err(&self) -> Result<&dyn Socket, Error>

pub fn as_inode_handle_or_err(&self) -> Result<&InodeHandle, Error>