aster_kernel::fs

Module file_table

Macros

get_file_fast 🔒

Gets a file from a file descriptor as fast as possible.

Structs

FdFlags

FileTable

FileTableEntry

OwnerObserver 🔒

Traits

WithFileTable

A helper trait that provides methods to operate the file table.

Type Aliases

FileDesc

Owner 🔒