aster_kernel::fs::file_table

Macro get_file_fast

macro_rules! get_file_fast {
    ($file_table:expr, $file_desc:expr) => { ... };
}

Gets a file from a file descriptor as fast as possible.

file_table should be a mutable borrow of the file table contained in the file_table field (which is a RefCell) in ThreadLocal. A mutable borrow is required because its exclusivity can be useful for achieving lockless file lookups.

If the file table is not shared with another thread, this macro will be free of locks (RwArc::read) and free of reference counting (Arc::clone).

If the file table is shared, the read lock is taken, the file is cloned, and then the read lock is released. Cloning and releasing the lock is necessary because we cannot hold such locks when operating on files, since many operations on files can block.

Note: This has to be a macro due to a limitation in the Rust borrow check implementation. Once https://github.com/rust-lang/rust/issues/58910 is fixed, we can try to convert this macro to a function.