aster_kernel::fs

Module fs_impls

Concrete file system implementations.

This module contains all the specific file system implementations supported by the kernel.

Modules

cgroupfs

configfs

devpts

exfat

ext2

A safe Rust Ext2 filesystem.

overlayfs

procfs

pseudofs

The infrastructure of pseudo FSes along with some naive implementations of pseudo FSes.

ramfs

Ramfs based on PageCache

sysfs

tmpfs

Temporary file system (tmpfs) based on ramfs.

Functions

init 🔒

init_on_each_cpu 🔒