aster_kernel::fs::fs_impls::ext2::super_block

Struct SuperBlock

pub(super) struct SuperBlock {

    inodes_count: u32,
    blocks_count: u32,
    reserved_blocks_count: u32,
    free_blocks_count: u32,
    free_inodes_count: u32,
    first_data_block: u32,
    block_size: usize,
    frag_size: usize,
    nr_blocks_per_group: u32,
    nr_frags_per_group: u32,
    nr_inodes_per_group: u32,
    nr_inode_table_blocks_per_group: u32,
    mtime: Duration,
    wtime: Duration,
    mnt_count: u16,
    max_mnt_count: u16,
    magic: u16,
    state: FsState,
    errors_behavior: ErrorsBehavior,
    last_check_time: Duration,
    check_interval: Duration,
    creator_os: OsId,
    rev_level: RevLevel,
    default_reserved_uid: u32,
    default_reserved_gid: u32,
    first_ino: u32,
    inode_size: usize,
    block_group_idx: usize,
    feature_compat: FeatureCompatSet,
    feature_incompat: FeatureInCompatSet,
    feature_ro_compat: FeatureRoCompatSet,
    uuid: [u8; 16],
    volume_name: FixedStr<16>,
    last_mounted_dir: FixedStr<64>,
    prealloc_file_blocks: u8,
    prealloc_dir_blocks: u8,
    min_rev_level: u16,
    algorithm_usage_bitmap: u32,
    padding1: u16,
    journal_uuid: [u8; 16],
    journal_ino: u32,
    journal_dev: u32,
    last_orphan: u32,
    hash_seed: [u32; 4],
    def_hash_version: u8,
    reserved_char_pad: u8,
    reserved_word_pad: u16,
    default_mount_opts: u32,
    first_meta_bg: u32,
    reserved: Reserved,

}

Validated, Rust-typed in-memory representation of the ext2 superblock.

Fields
inodes_count: u32

Total number of inodes.

blocks_count: u32

Total number of blocks.

reserved_blocks_count: u32

Total number of reserved blocks.

free_blocks_count: u32

Total number of free blocks.

free_inodes_count: u32

Total number of free inodes.

first_data_block: u32

First data block.

block_size: usize

Block size.

frag_size: usize

Fragment size.

nr_blocks_per_group: u32

Number of blocks in each block group.

nr_frags_per_group: u32

Number of fragments in each block group.

nr_inodes_per_group: u32

Number of inodes in each block group.

nr_inode_table_blocks_per_group: u32

Number of inode table blocks in each group.

mtime: Duration

Mount time.

wtime: Duration

Write time.

mnt_count: u16

Mount count.

max_mnt_count: u16

Maximal mount count.

magic: u16

Magic signature.

state: FsState

File system state.

errors_behavior: ErrorsBehavior

Behavior when detecting errors.

last_check_time: Duration

Time of last check.

check_interval: Duration

Interval between checks.

creator_os: OsId

Creator OS ID.

rev_level: RevLevel

Revision level.

default_reserved_uid: u32

Default UID for reserved blocks.

default_reserved_gid: u32

Default GID for reserved blocks.

first_ino: u32

First non-reserved inode number.

inode_size: usize

Size of inode structure.

block_group_idx: usize

Block group that this superblock is part of (if backup copy).

feature_compat: FeatureCompatSet

Compatible feature set.

feature_incompat: FeatureInCompatSet

Incompatible feature set.

feature_ro_compat: FeatureRoCompatSet

Read-only-compatible feature set.

uuid: [u8; 16]

128-bit UUID for the volume.

volume_name: FixedStr<16>

Volume name.

last_mounted_dir: FixedStr<64>

Directory where last mounted.

prealloc_file_blocks: u8

Number of blocks to preallocate for files.

prealloc_dir_blocks: u8

Number of blocks to preallocate for directories.

min_rev_level: u16algorithm_usage_bitmap: u32padding1: u16journal_uuid: [u8; 16]journal_ino: u32journal_dev: u32last_orphan: u32hash_seed: [u32; 4]def_hash_version: u8reserved_char_pad: u8reserved_word_pad: u16default_mount_opts: u32first_meta_bg: u32reserved: Reserved
Implementations

impl SuperBlock

pub(super) const fn block_size(&self) -> usize

Returns the block size.

pub(super) fn max_file_size(&self) -> usize

Returns the maximum regular file size supported by this ext2 instance.

const fn max_blocks(&self) -> u64

pub(super) const fn inode_size(&self) -> usize

Returns the size of inode structure.

pub(super) const fn fragment_size(&self) -> usize

Returns the fragment size.

pub(super) const fn total_inodes(&self) -> u32

Returns total number of inodes.

pub(super) const fn total_blocks(&self) -> u32

Returns total number of blocks.

pub(super) const fn nr_blocks_per_group(&self) -> u32

Returns the number of blocks in each block group.

pub(super) const fn group_first_block_no(&self, group_idx: usize) -> u32

Returns the first block number of a block group.

pub(super) fn group_last_block_no(&self, group_idx: usize) -> u32

Returns the last block number of a block group.

pub(super) fn is_data_block_valid(&self, start_blk: u32, count: u32) -> bool

Returns whether a data block range is valid.

pub(super) const fn first_data_block(&self) -> u32

Returns the first data block number.

pub(super) const fn nr_inodes_per_group(&self) -> u32

Returns the number of inodes in each block group.

pub(super) const fn nr_inode_table_blocks_per_group(&self) -> u32

Returns the number of inode table blocks in each block group.

pub(super) const fn first_ino(&self) -> u32

Returns the first non-reserved inode number.

pub(super) const fn nr_block_groups(&self) -> u32

Returns the number of block groups.

pub(super) const fn group_descriptor_blocks_count(&self) -> u32

Returns the number of group descriptor blocks in each superblock copy.

pub(super) const fn free_blocks_count(&self) -> u32

Returns the number of free blocks.

pub(super) const fn reserved_blocks_count(&self) -> u32

Returns the number of reserved blocks.

pub(super) const fn default_reserved_uid(&self) -> u32

Returns the default UID for reserved blocks.

pub(super) const fn default_reserved_gid(&self) -> u32

Returns the default GID for reserved blocks.

pub(super) fn inc_free_blocks(&mut self, count: u32) -> Result<(), Error>

Increases the number of free blocks.

pub(super) const fn set_free_blocks_count(&mut self, count: u32)

Overwrites the free blocks counter with a recomputed value.

pub(super) fn dec_free_blocks(&mut self, count: u32) -> Result<(), Error>

Decreases the number of free blocks.

pub(super) const fn free_inodes_count(&self) -> u32

Returns the number of free inodes.

pub(super) const fn set_free_inodes_count(&mut self, count: u32)

Overwrites the free inodes counter with a recomputed value.

pub(super) fn inc_free_inodes(&mut self) -> Result<(), Error>

Increases the number of free inodes.

pub(super) const fn set_wtime(&mut self, time: Duration)

Sets the last write time.

pub(super) fn dec_free_inodes(&mut self) -> Result<(), Error>

Decreases the number of free inodes.

pub(super) fn is_backup_group(&self, block_group_idx: usize) -> bool

Checks if the block group will backup the super block.

pub(super) fn has_super_block(&self, block_group_idx: usize) -> bool

Returns whether the given group stores the primary or a backup superblock.

pub(super) fn total_metadata_blocks(&self) -> u32

Computes the metadata overhead subtracted from reported block totals.

pub(super) fn bid(&self, block_group_idx: usize) -> u32

Returns the starting block ID of the superblock copy inside the block group identified by block_group_idx.

pub(super) fn group_descriptors_bid(&self, block_group_idx: usize) -> u32

Returns the starting block ID of the block-group descriptor table inside the block group identified by block_group_idx.

const fn state(&self) -> FsState

const fn rev_level(&self) -> RevLevel

const fn feature_compat(&self) -> FeatureCompatSet

const fn feature_incompat(&self) -> FeatureInCompatSet

const fn feature_ro_compat(&self) -> FeatureRoCompatSet