aster_kernel::fs::fs_impls

Module pseudofs

The infrastructure of pseudo FSes along with some naive implementations of pseudo FSes.

Pseudo FSes are those that have no physical storage. Some pseudo FSes are trivial, e.g., PipeFs, SockFs, PidfdFs, and AnonInodeFs. Such trivial pseudo FSes are directly implemented in this module using a generic pseudo FS implementation called NaivePseudoFs.

On the other hand, some pseudo FSes are quite complex, including RamFs, ProcFs, and SysFs. Just to name a few. They are implemented in their own modules.

Pseudo FSes are not backed by physical storage devices, but they still need to be assigned container device IDs (see Metadata::container_dev_id). Linux’s behavior is to assign anonymous device IDs (whose major IDs are 0). As such, this module provides AnonDeviceId to acquire and recycle anonymous device IDs.

Re-exports

pub use allocator::AnonDeviceId;

pub use anon_inodefs::AnonInodeFs;

pub use nsfs::NsCommonOps;

pub use nsfs::NsFile;

pub use nsfs::NsType;

pub use nsfs::StashedDentry;

pub use pidfdfs::PidfdFs;

pub use sockfs::SockFs;

Modules

allocator 🔒

anon_inodefs 🔒

nsfs 🔒

pidfdfs 🔒

pipefs 🔒

sockfs 🔒

Structs

NaivePseudoFs

A pseudo file system that manages pseudo inodes, such as pipe inodes and socket inodes.

PseudoInode

A pseudo inode that does not correspond to any real path in the file system.

Enums

PseudoInodeType

Constants

ROOT_INO 🔒

Root Inode ID.

Functions

init 🔒

now 🔒