aster_kernel::fs::inode_handle

Trait FileIo

pub trait FileIo:
    Pollable
    + InodeIo
    + Any
    + Send
    + Sync
    + 'static {
    // Required methods
    fn check_seekable(&self) -> Result<(), Error>;
    fn is_offset_aware(&self) -> bool;

    // Provided methods
    fn mappable(&self) -> Result<Mappable, Error> { ... }
    fn ioctl(&self, _raw_ioctl: RawIoctl) -> Result<i32, Error> { ... }
}

A trait for file-like objects that provide custom I/O operations.

This trait is typically implemented for special files like devices or named pipes (FIFOs), which have behaviors different from regular on-disk files.

Required Methods

fn check_seekable(&self) -> Result<(), Error>

Checks whether the seek() operation should fail.

fn is_offset_aware(&self) -> bool

Returns whether the read()/write() operation should use and advance the offset.

If FileIo::check_seekable succeeds but this method returns false, the offset in the seek() operation will be ignored. In that case, the seek() operation will do nothing but succeed.

Provided Methods

fn mappable(&self) -> Result<Mappable, Error>

fn ioctl(&self, _raw_ioctl: RawIoctl) -> Result<i32, Error>