aster_kernel::fs

Module notify

Modules

inotify

Structs

AtomicFsEvents 🔒

FsEventPublisher

Publishes filesystem events to subscribers.

FsEvents

Represents filesystem events that have occurred.

Traits

FsEventSubscriber

Represents a subscriber to filesystem events on an FsEventPublisher.

Functions

notify_inode 🔒

Sends a filesystem notification event to all subscribers of an inode.

notify_inode_with_name 🔒

Sends a filesystem notification event with a name to all subscribers of an inode.

notify_parent 🔒

Notifies a path’s parent and the path itself about filesystem events.

on_access

Notifies that a file was accessed.

on_attr_change

Notifies that a file’s attributes changed.

on_change

Notifies that a path’s content was changed.

on_close

Notifies that a file was closed.

on_create

Notifies that a file was created.

on_delete

Notifies that a file was deleted from a directory.

on_inode_removed

Notifies that an inode was removed (link count reached 0).

on_link

Notifies that a file was linked to a directory.

on_link_count

Notifies that an inode’s link count changed.

on_mkdir

Notifies that a directory was created.

on_modify

Notifies that a file was modified.

on_open

Notifies that a file was opened.