aster_kernel::fs::notify

Trait FsEventSubscriber

pub trait FsEventSubscriber:
    Any
    + Send
    + Sync {
    // Required methods
    fn deliver_event(&self, events: FsEvents, name: Option<String>) -> bool;
    fn interesting_events(&self) -> FsEvents;

    // Provided method
    fn is_oneshot_and_dead(&self) -> bool { ... }
}

Represents a subscriber to filesystem events on an FsEventPublisher.

A subscriber receives notifications from a publisher when filesystem events occur that match the subscriber’s interesting events. The subscriber specifies which events it is interested in using FsEvents, which define the types of events (e.g., read, write, modify, delete) the subscriber wants to be notified about. When an event occurs, the publisher (attached to an inode) broadcasts it to all subscribers whose interesting events match the event type.

Required Methods

fn deliver_event(&self, events: FsEvents, name: Option<String>) -> bool

Delivers a filesystem event notification to the subscriber.

Returns whether the subscriber is a one-shot subscriber and the event has been delivered. If there are no one-shot subscribers, simply return false here. Otherwise, Self::is_oneshot_and_dead should be implemented correspondingly.

Invariant: This method must not sleep or perform blocking operations. The publisher may hold a spin lock when calling this method.

fn interesting_events(&self) -> FsEvents

Returns the events that this subscriber is interested in.

Provided Methods

fn is_oneshot_and_dead(&self) -> bool

Returns whether the subscriber is a one-shot subscriber and an event has been delivered.

This method should return true if and only if a previous call to Self::deliver_event has already returned true.