aster_kernel::fs::overlayfs

Module fs

Structs

OverlayConfig

Holds various mode settings and feature toggles.

OverlayDirVisitor 🔒

A visitor used by OverlayFs that merges the objects from the upper layer and the lower layer.

OverlayFs

An OverlayFs is a union pseudo file system employed to merge upper and lower directories that potentially comes from different file systems, into a single, unified view at a designated mount point.

OverlayFsType 🔒

OverlayInode 🔒

Provides an unified inode abstraction for its user, internal it manages the layered regular inodes.

OverlayLower 🔒

The immutable lower layer of an OverlayFs. A lower layer may contain multiple Paths with different mount points.

OverlaySB 🔒

OverlayUpper 🔒

The mutable upper layer of an OverlayFs.

OverlayWork 🔒

The work directory. Must reside in the same file system as the upper layer.

UniqueNoGenerator 🔒

Constants

OPAQUE_DIR_XATTR_NAME 🔒

OVERLAY_FS_MAGIC 🔒

WHITEOUT_AND_OPAQUE_XATTR_VALUE 🔒

WHITEOUT_PREFIX 🔒

WHITEOUT_PREFIX_SIZE 🔒

WHITEOUT_XATTR_NAME 🔒

Functions

is_opaque_dir 🔒

whiteout_name 🔒

Type Aliases

LayerIdx 🔒

The index of the layer of an OverlayFs.