aster_kernel::fs::path::dentry

Struct DirDentry

pub(super) struct DirDentry<'a> {
    inner: &'a Dentry,
    children: &'a RwMutex<DentryChildren>,
}

A Dentry wrapper that has been validated to represent a directory.

Fields
inner: &'a Dentrychildren: &'a RwMutex<DentryChildren>
Implementations

impl DirDentry<'_>

pub(super) fn create( &self, name: &str, type_: InodeType, mode: InodeMode, ) -> Result<Arc<Dentry>, Error>

Creates a Dentry by creating a new inode of the type_ with the mode.

pub(super) fn lookup_via_cache( &self, name: &str, ) -> Result<Option<Arc<Dentry>>, Error>

Lookups a target Dentry from the cache in children.

pub(super) fn lookup_via_fs(&self, name: &str) -> Result<Arc<Dentry>, Error>

Lookups a target Dentry from the file system.

pub(super) fn mknod( &self, name: &str, mode: InodeMode, type_: MknodType, ) -> Result<Arc<Dentry>, Error>

Creates a Dentry by making an inode of the type_ with the mode.

pub(super) fn link( &self, old_inode: &Arc<dyn Inode>, name: &str, ) -> Result<(), Error>

Links a new Dentry by link() the old inode.

pub(super) fn unlink(&self, name: &str) -> Result<(), Error>

Deletes a Dentry by unlink() the inner inode.

pub(super) fn rmdir(&self, name: &str) -> Result<(), Error>

Deletes a directory Dentry by rmdir() the inner inode.

pub(super) fn rename( old_dir_arc: &Arc<Dentry>, old_name: &str, new_dir_arc: &Arc<Dentry>, new_name: &str, ) -> Result<(), Error>

Renames a Dentry to the new Dentry by rename() the inner inode.