aster_kernel::fs::path::mount

Struct Mount

pub struct Mount {
    id: usize,
    root_dentry: Arc<Dentry>,
    mountpoint: RwLock<Option<Arc<Dentry>>>,
    fs: Arc<dyn FileSystem>,
    source: Option<String>,
    parent: RwLock<Option<Weak<Mount>>>,
    pub(super) children: RwLock<HashMap<DentryKey, Arc<Self>>>,
    mnt_ns: Weak<MountNamespace>,
    propagation: RwLock<MountPropType>,
    flags: AtomicPerMountFlags,
    this: Weak<Self>,
}

A Mount represents a mounted filesystem instance in the VFS.

Each Mount can be viewed as a node in the mount tree, maintaining mount-related information and the structure of the mount tree.

Fields
id: usize

Global unique identifier for the mount node.

root_dentry: Arc<Dentry>

Root dentry.

mountpoint: RwLock<Option<Arc<Dentry>>>

Mountpoint dentry. A mount node can be mounted on one dentry of another mount node, which makes the mount being the child of the mount node.

fs: Arc<dyn FileSystem>

The associated FS.

source: Option<String>

The mount source (e.g., a device path like “/dev/vda” or a filesystem name like “proc”).

The source is stored in Mount instead of requiring each filesystem to provide it. If a filesystem does not provide a source, it falls back to the value stored in Mount. This behavior aligns with that of Linux. Concrete examples can be found here: https://github.com/asterinas/asterinas/pull/2929#discussion_r2729739818.

Reference: https://elixir.bootlin.com/linux/v6.17/source/fs/mount.h#L68

parent: RwLock<Option<Weak<Mount>>>

The parent mount node.

children: RwLock<HashMap<DentryKey, Arc<Self>>>

Child mount nodes which are mounted on one dentry of self.

mnt_ns: Weak<MountNamespace>

The associated mount namespace.

propagation: RwLock<MountPropType>

Propagation type of this mount (e.g., private, shared).

flags: AtomicPerMountFlags

The flags of this mount.

this: Weak<Self>

Reference to self.

Implementations

impl Mount

pub(in fs) fn new_root( fs: Arc<dyn FileSystem>, mnt_ns: Weak<MountNamespace>, ) -> Arc<Self>

Creates a root mount node with an associated FS.

The root mount node is not mounted on other mount nodes (which means it has no parent). The root inode of the fs will form the inner root dentry.

It is allowed to create a mount node even if the fs has been provided to another mount node. It is the fs’s responsibility to ensure the data consistency.

pub(in fs) fn new_pseudo( fs: Arc<dyn FileSystem>, ) -> Arc<Self>

Creates a pseudo mount node with an associated FS.

This pseudo mount is not mounted on other mount nodes, has no parent, and does not belong to any mount namespace.

fn new( fs: Arc<dyn FileSystem>, flags: PerMountFlags, parent_mount: Option<Weak<Mount>>, mnt_ns: Weak<MountNamespace>, source: Option<String>, ) -> Arc<Self>

The internal constructor.

A root mount node has no mountpoint, while other mount nodes must have one.

Here, a Mount is instantiated without an initial mountpoint, avoiding fixed mountpoint limitations. This allows the root mount node to exist without a mountpoint, ensuring uniformity and security, while all other mount nodes must be explicitly assigned a mountpoint to maintain structural integrity.

pub fn id(&self) -> usize

Gets the mount ID.

pub(in fs) fn source( &self, ) -> Option<&str>

Returns the mount source.

pub(super) fn do_mount( self: &Arc<Self>, fs: Arc<dyn FileSystem>, flags: PerMountFlags, mountpoint: &Arc<Dentry>, source: Option<String>, ) -> Result<Arc<Self>, Error>

Mounts a fs on the mountpoint, it will create a new child mount node.

If the given mountpoint has already been mounted, then its mounted child mount node will be updated.

The mountpoint should belong to this mount node, or an error is returned.

It is allowed to mount a fs even if the fs has been provided to another mountpoint. It is the fs’s responsibility to ensure the data consistency.

If the source is provided by user, it will be recorded in the new mount.

Return the mounted child mount.

pub(super) fn do_unmount(&self, mountpoint: &Dentry) -> Result<Arc<Self>, Error>

Unmounts a child mount node from the mountpoint and returns it.

The mountpoint should belong to this mount node, or an error is returned.

fn clone_mount( &self, root_dentry: &Arc<Dentry>, new_ns: Option<&Weak<MountNamespace>>, ) -> Arc<Self>

Clones a mount node with the an root Dentry.

The new mount node will have the same fs as the original one and have no parent and children. We should set the parent and children manually.

If the new_ns is set, the new mount will belong to the given mount namespace. Otherwise, it will belong to the same mount namespace as the current mount.

pub(super) fn clone_mount_tree( &self, root_dentry: &Arc<Dentry>, new_ns: Option<&Weak<MountNamespace>>, recursive: bool, ) -> Arc<Self>

Clones a mount tree starting from the specified root Dentry.

The new mount tree will replicate the structure of the original tree. The new tree is a separate entity rooted at the given Dentry, and the original tree remains unchanged.

If recursive is set to true, the entire tree will be copied. Otherwise, only the root mount node will be copied.

If the new_ns is set, the new mount tree will belong to the given mount namespace. Otherwise, it will belong to the same mount namespace as the current mount.

pub(super) fn set_propagation(&self, prop: MountPropType, recursive: bool)

Sets the propagation type of this mount.

pub(super) fn detach_from_parent(&self)

Detaches the mount node from the parent mount node.

fn attach_to_path(&self, target_path: &Path)

Attaches the mount node to the mountpoint.

pub(super) fn graft_mount_tree(&self, target_path: &Path) -> Result<(), Error>

Grafts the mount node tree to the mountpoint.

pub(super) fn get(&self, mountpoint: &Dentry) -> Option<Arc<Self>>

Gets a child mount node from the mountpoint if any.

pub(in fs) fn root_dentry( &self, ) -> &Arc<Dentry>

Gets the root Dentry of this mount node.

pub(in fs) fn mountpoint( &self, ) -> Option<Arc<Dentry>>

Gets the mountpoint Dentry of this mount node if any.

pub(super) fn set_mountpoint(&self, dentry: &Arc<Dentry>)

Sets the mountpoint.

pub(super) fn clear_mountpoint(&self)

Clears the mountpoint.

pub(super) fn sync(&self) -> Result<(), Error>

Flushes all pending filesystem metadata and cached file data to the device.

pub(super) fn remount( &self, mount_flags: PerMountFlags, fs_flags: Option<FsFlags>, data: Option<CString>, ctx: &Context<'_>, ) -> Result<(), Error>

pub(in fs) fn parent( &self, ) -> Option<Weak<Self>>

Gets the parent mount node if any.

pub(super) fn mnt_ns(&self) -> &Weak<MountNamespace>

Gets the associated mount namespace.

pub(in fs) fn fs( &self, ) -> &Arc<dyn FileSystem>

Gets the associated FS.

pub(in fs) fn flags( &self, ) -> PerMountFlags

pub(super) fn set_parent(&self, mount: Option<&Arc<Mount>>)

Sets the parent mount node.

In some cases we may need to reset the parent of the created Mount, such as move mount.

pub(super) fn find_corresponding_mount( &self, mnt_ns: &Arc<MountNamespace>, ) -> Option<Arc<Self>>

Finds the corresponding Mount in the given mount namespace.

fn this(&self) -> Arc<Self>