aster_kernel::fs::path

Struct Path

pub struct Path {
    mount: Arc<Mount>,
    dentry: Arc<Dentry>,
}

A Path is used to represent an exact location in the VFS tree.

Each Path corresponds to a node in the VFS tree, and a single node may have multiple Path instances referencing it due to mount operations.

Fields
mount: Arc<Mount>dentry: Arc<Dentry>
Implementations

impl Path

pub fn new_fs_root(mount: Arc<Mount>) -> Self

Creates a new Path to represent the root directory of a file system.

pub fn new_fs_child( &self, name: &str, type_: InodeType, mode: InodeMode, ) -> Result<Self, Error>

Creates a new Path to represent the child directory of a file system.

pub(super) fn new_pseudo( mount: Arc<Mount>, inode: Arc<dyn Inode>, name_fn: fn(&dyn Inode) -> String, ) -> Self

Creates a new pseudo Path.

pub(super) fn new(mount: Arc<Mount>, dentry: Arc<Dentry>) -> Self

pub fn mount_node(&self) -> &Arc<Mount>

Gets the mount node of current Path.

pub(super) fn dentry(&self) -> &Arc<Dentry>

Gets the dentry of current Path.

pub fn is_mount_root(&self) -> bool

Returns true if the current Path is the root of its mount.

pub fn open(&self, open_args: OpenArgs) -> Result<InodeHandle, Error>

Opens the Path with the given OpenArgs.

Returns an InodeHandle on success.

pub fn name(&self) -> String

Gets the real name of the Path from its dentry.

pub(super) fn parent_within_mount(&self) -> Option<Self>

Gets the parent Path within the same mount.

This method returns the parent path within the same filesystem/mount. It does NOT cross mount boundaries. If the current path is the root of a mount, it will return None.

For cross-filesystem parent lookup, use effective_parent() instead.

pub(super) fn get_top_path(self) -> Self

Gets the top Path of the current.

Used when different file systems are mounted on the same mount point.

For example, first mount /dev/sda1 /mnt and then mount /dev/sda2 /mnt. After the second mount is completed, the content of the first mount will be overridden. We need to recursively obtain the top Path.

pub(super) fn find_corresponding_mount( &self, mnt_ns: &Arc<MountNamespace>, ) -> Option<Self>

Finds the corresponding Path in the given mount namespace.

fn is_pseudo(&self) -> bool

Returns true if the Path represents a pseudo file.

fn this(&self) -> Self

impl Path

pub fn mount( &self, fs: Arc<dyn FileSystem>, flags: PerMountFlags, source: Option<String>, ctx: &Context<'_>, ) -> Result<Arc<Mount>, Error>

Mounts a filesystem at the current path.

This method attaches a given filesystem to the VFS tree at the location represented by self. The current path becomes the mountpoint for the new filesystem.

Returns the newly created child mount on success.

Errors

Returns ENOTDIR if the path is not a directory. Returns EINVAL if attempting to mount on root or if the path is not in the current mount namespace.

pub fn unmount(&self, ctx: &Context<'_>) -> Result<Arc<Mount>, Error>

Unmounts the filesystem mounted at the current path.

Returns the unmounted child mount on success.

Errors

Returns EINVAL in the following cases:

The current path is not a mount root.

The mount of the current path is the root mount.

The current path is not in the current mount namespace.

pub fn remount( &self, mount_flags: PerMountFlags, fs_flags: Option<FsFlags>, data: Option<CString>, ctx: &Context<'_>, ) -> Result<(), Error>

Remounts the filesystem with new PerMountFlags and optionally new FsFlags.

If fs_flags is provided, it will update the flags of the mounted filesystem, otherwise, only the flags of the current mount will be updated.

Errors

Returns EINVAL in the following cases:

The current path is not a mount root.

The current path is not in the current mount namespace.

pub fn bind_mount_to( &self, dst_path: &Self, recursive: bool, ctx: &Context<'_>, ) -> Result<(), Error>

Creates a bind mount from the current path to the destination path.

Creates a new mount tree that mirrors either the root mount (non-recursive) or the entire mount subtree (recursive), and attaches it to the destination path.

Errors

Returns ENOTDIR if the dst_path is not a directory. Returns EINVAL if either source or destination path is not in the current mount namespace.

pub fn move_mount_to( &self, dst_path: &Self, ctx: &Context<'_>, ) -> Result<(), Error>

Moves a mount tree from the current path to the destination path.

Errors

Returns ENOTDIR if the dst_path is not a directory. Returns EINVAL in the following cases:

The current path is not a mount root.

The mount of the current path is the root mount.

Either source or destination path is not in the current mount namespace

pub fn set_mount_propagation( &self, prop: MountPropType, recursive: bool, ctx: &Context<'_>, ) -> Result<(), Error>

Sets the propagation type of the mount of this Path.

impl Path

pub fn inode(&self) -> &Arc<dyn Inode>

pub fn type_(&self) -> InodeType

pub fn mknod( &self, name: &str, mode: InodeMode, type_: MknodType, ) -> Result<Self, Error>

Creates a Path by making an inode of the type_ with the mode.

pub fn link(&self, old: &Self, name: &str) -> Result<(), Error>

Links a new name for the Path.

pub fn unlink(&self, name: &str) -> Result<(), Error>

Unlinks a name from the Path.

pub fn rmdir(&self, name: &str) -> Result<(), Error>

Removes a directory by rmdir() the inner inode.

pub fn rename( &self, old_name: &str, new_dir: &Self, new_name: &str, ) -> Result<(), Error>

Renames a Path to the new Path by rename() the inner inode.

impl Path

pub fn fs(&self) -> Arc<dyn FileSystem>

pub fn sync_all(&self) -> Result<(), Error>

pub fn sync_data(&self) -> Result<(), Error>

pub fn metadata(&self) -> Metadata

pub fn mode(&self) -> Result<InodeMode, Error>

pub fn set_mode(&self, mode: InodeMode) -> Result<(), Error>

pub fn size(&self) -> usize

pub fn resize(&self, size: usize) -> Result<(), Error>

pub fn owner(&self) -> Result<Uid, Error>

pub fn set_owner(&self, uid: Uid) -> Result<(), Error>

pub fn group(&self) -> Result<Gid, Error>

pub fn set_group(&self, gid: Gid) -> Result<(), Error>

pub fn atime(&self) -> Duration

pub fn set_atime(&self, time: Duration)

pub fn mtime(&self) -> Duration

pub fn set_mtime(&self, time: Duration)

pub fn ctime(&self) -> Duration

pub fn set_ctime(&self, time: Duration)

pub fn set_xattr( &self, name: XattrName<'_>, value_reader: &mut VmReader<'_>, flags: XattrSetFlags, ) -> Result<(), Error>

pub fn get_xattr( &self, name: XattrName<'_>, value_writer: &mut VmWriter<'_>, ) -> Result<usize, Error>

pub fn list_xattr( &self, namespace: XattrNamespace, list_writer: &mut VmWriter<'_>, ) -> Result<usize, Error>

pub fn remove_xattr(&self, name: XattrName<'_>) -> Result<(), Error>