aster_kernel::fs

Module procfs

Modules

cmdline 🔒

This module offers /proc/cmdline file support, which provides information about arguments passed to the kernel at boot time.

cpuinfo 🔒

This module offers /proc/cpuinfo file support, which provides information about the CPU architecture, cores, and other details.

filesystems 🔒

loadavg 🔒

This module offers /proc/loadavg file support, which tells the user space about the cpu load average for the last 1, 5, and 15 minutes.

meminfo 🔒

This module offers /proc/meminfo file support, which tells the user space about the memory statistics in the entire system. The definition of the fields are similar to that of Linux’s but there exist differences.

mounts 🔒

pid 🔒

self_ 🔒

stat 🔒

This module offers /proc/stat file support, which provides information about kernel system statistics.

sys 🔒

template 🔒

thread_self 🔒

uptime 🔒

This module offers /proc/uptime file support, which provides information about the system uptime and idle time.

version 🔒

This module offers /proc/version file support, which provides information about the kernel version.

Structs

ProcFs 🔒

ProcFsType 🔒

RootDirOps 🔒

Represents the inode at /proc.

Constants

BLOCK_SIZE 🔒

Block size.

PROC_MAGIC 🔒

Magic number.

PROC_ROOT_INO 🔒

Root Inode ID.

Functions

init 🔒

init_on_each_cpu 🔒