aster_kernel::fs

Module pseudofs

Structs

AnonInodeFs

PidfdFs

PipeFs 🔒

PipeFsType 🔒

PseudoFs

A pseudo file system that manages pseudo inodes, such as pipe inodes and socket inodes.

PseudoInode

A pseudo inode that does not correspond to any real path in the file system.

SockFs

SockFsType 🔒

Enums

PseudoInodeType

Constants

ANON_INODEFS_MAGIC 🔒

PIDFDFS_MAGIC 🔒

PIPEFS_MAGIC 🔒

ROOT_INO 🔒

Root Inode ID.

SOCKFS_MAGIC 🔒

Functions

init 🔒

now 🔒