aster_kernel::fs::utils::direntry_vec

Trait DirEntryVecExt

pub trait DirEntryVecExt {
    // Required methods
    fn find_entry_by_name(&self, name: &str) -> Option<&Arc<dyn Inode>>;
    fn put_entry_if_not_found(
        &mut self,
        name: &str,
        f: impl FnOnce() -> Arc<dyn Inode>,
    ) -> &Arc<dyn Inode>;
    fn remove_entry_by_name(
        &mut self,
        name: &str,
    ) -> Option<(String, Arc<dyn Inode>)>;
}

Required Methods

fn find_entry_by_name(&self, name: &str) -> Option<&Arc<dyn Inode>>

Finds the entry by the name.

fn put_entry_if_not_found( &mut self, name: &str, f: impl FnOnce() -> Arc<dyn Inode>, ) -> &Arc<dyn Inode>

Puts the entry given by f into the vector if it is not found by the name.

fn remove_entry_by_name( &mut self, name: &str, ) -> Option<(String, Arc<dyn Inode>)>

Removes the entry by the name.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.