aster_kernel::fs::utils::inode_mode

Macro chmod

macro_rules! chmod {
    ($mode:expr) => { ... };
    ($mode:expr, $who:ident + $perms:ident $(, $($rest:tt)*)?) => { ... };
    ($mode:expr, $who:ident - $perms:ident $(, $($rest:tt)*)?) => { ... };
    ($mode:expr, $who:ident = $perms:ident $(, $($rest:tt)*)?) => { ... };
    ($mode:expr, $who:ident = $(, $($rest:tt)*)?) => { ... };
    (@apply $mode:expr, $who:ident, $op:expr, $perms:ident $(, $($rest:tt)*)?) => { ... };
}

Changes an inode mode.

Syntax

If you are familiar with the chmod tool, the syntax of the chmod macro should appear natural and self-explaining.

The chmod macro takes two or more arguments. The first argument is an initial value of the inode mode. The rest arguments are one or more symbolic mode, each of which dictates how the current value of the inode mode shall be changed.

The format of a symbolic mode is as follows:

[ugoa][+-=][perms]

where [perms] is either zero or more letters from rwx.

The user part, [ugoa], controls the target users whose access to the file will be changed:

u: the user who owns it;

g: other users in the file owner’s group;

o: other users not in the file owner’s group;

a: all users.

The operator part, [+-=], indicates how the file permission bits are changed:

+: [perms] are added to the target users’ permissions;

-: [perms] are removed from the target users’ permissions;

=: [perms] are set to the target users’ permissions.

The permission part, [perms], is a combination of rwx.

r: the read permission;

w: the write permission;

x: the execute/search permission.

Examples

// Add the read and write permissions to everyone
let mode0 = mkmod!(a+rw);
assert!(mode0.bits() == 0o666);

// Add the execute permissions to the owner user
let mode1 = chmod!(mode0, u+x);
assert!(mode1.bits() == 0o766);

// Combine the above two steps into one invocation
let mode2 = mkmod!(a+rw, u+x);
assert!(mode2.bits() == 0o766);