aster_kernel::fs::utils::page_cache

Trait CachePageExt

pub trait CachePageExt {
    // Required method
    fn metadata(&self) -> &CachePageMeta;

    // Provided methods
    fn alloc_uninit() -> Result<Frame<CachePageMeta>, Error> { ... }
    fn alloc_zero(state: PageState) -> Result<Frame<CachePageMeta>, Error> { ... }
    fn load_state(&self) -> PageState { ... }
    fn store_state(&mut self, new_state: PageState) { ... }
}

Required Methods

fn metadata(&self) -> &CachePageMeta

Gets the metadata associated with the cache page.

Provided Methods

fn alloc_uninit() -> Result<Frame<CachePageMeta>, Error>

Allocates a new cache page which content and state are uninitialized.

fn alloc_zero(state: PageState) -> Result<Frame<CachePageMeta>, Error>

Allocates a new zeroed cache page with the wanted state.

fn load_state(&self) -> PageState

Loads the current state of the cache page.

fn store_state(&mut self, new_state: PageState)

Stores a new state for the cache page.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.