aster_kernel::fs::vfs::fs_apis

Module inode

Structs

Extension

An extension is a set of object groups that is attached to an inode.

InodeWriter

Metadata

RevalidationPolicy

Policy that controls when the VFS revalidates cached inode entries.

Enums

FallocMode

Represents the various operation modes for fallocate.

HardLinkability

Describes whether an inode may get new hard links.

MknodType

RenameMode

The behavior of rename operation.

SymbolicLink

A symbolic link.

Traits

FileOps

Basic file operations.

Inode

Functions

has_dac_override_capability 🔒