aster_kernel::fs::vfs::fs_apis::inode

Trait Inode

pub trait Inode:
    Any
    + InodeIo
    + Send
    + Sync {

    // Required methods
    fn size(&self) -> usize;
    fn resize(&self, new_size: usize) -> Result<(), Error>;
    fn metadata(&self) -> Metadata;
    fn ino(&self) -> u64;
    fn type_(&self) -> InodeType;
    fn mode(&self) -> Result<InodeMode, Error>;
    fn set_mode(&self, mode: InodeMode) -> Result<(), Error>;
    fn owner(&self) -> Result<Uid, Error>;
    fn set_owner(&self, uid: Uid) -> Result<(), Error>;
    fn group(&self) -> Result<Gid, Error>;
    fn set_group(&self, gid: Gid) -> Result<(), Error>;
    fn atime(&self) -> Duration;
    fn set_atime(&self, time: Duration);
    fn mtime(&self) -> Duration;
    fn set_mtime(&self, time: Duration);
    fn ctime(&self) -> Duration;
    fn set_ctime(&self, time: Duration);
    fn fs(&self) -> Arc<dyn FileSystem>;
    fn extension(&self) -> &Extension;

    // Provided methods
    fn page_cache(&self) -> Option<Arc<Vmo>> { ... }
    fn create(
        &self,
        name: &str,
        type_: InodeType,
        mode: InodeMode,
    ) -> Result<Arc<dyn Inode>, Error> { ... }
    fn mknod(
        &self,
        name: &str,
        mode: InodeMode,
        type_: MknodType,
    ) -> Result<Arc<dyn Inode>, Error> { ... }
    fn open(
        &self,
        access_mode: AccessMode,
        status_flags: StatusFlags,
    ) -> Option<Result<Box<dyn FileIo>, Error>> { ... }
    fn readdir_at(
        &self,
        offset: usize,
        visitor: &mut dyn DirentVisitor,
    ) -> Result<usize, Error> { ... }
    fn link(&self, old: &Arc<dyn Inode>, name: &str) -> Result<(), Error> { ... }
    fn unlink(&self, name: &str) -> Result<(), Error> { ... }
    fn rmdir(&self, name: &str) -> Result<(), Error> { ... }
    fn lookup(&self, name: &str) -> Result<Arc<dyn Inode>, Error> { ... }
    fn rename(
        &self,
        old_name: &str,
        target: &Arc<dyn Inode>,
        new_name: &str,
    ) -> Result<(), Error> { ... }
    fn read_link(&self) -> Result<SymbolicLink, Error> { ... }
    fn write_link(&self, target: &str) -> Result<(), Error> { ... }
    fn sync_all(&self) -> Result<(), Error> { ... }
    fn sync_data(&self) -> Result<(), Error> { ... }
    fn fallocate(
        &self,
        mode: FallocMode,
        offset: usize,
        len: usize,
    ) -> Result<(), Error> { ... }
    fn revalidation_policy(&self) -> RevalidationPolicy { ... }
    fn revalidate_exists(&self, _name: &str, _child: &dyn Inode) -> bool { ... }
    fn revalidate_absent(&self, _name: &str) -> bool { ... }
    fn seek_end(&self) -> Option<usize> { ... }
    fn set_xattr(
        &self,
        name: XattrName<'_>,
        value_reader: &mut VmReader<'_>,
        flags: XattrSetFlags,
    ) -> Result<(), Error> { ... }
    fn get_xattr(
        &self,
        name: XattrName<'_>,
        value_writer: &mut VmWriter<'_>,
    ) -> Result<usize, Error> { ... }
    fn list_xattr(
        &self,
        namespace: XattrNamespace,
        list_writer: &mut VmWriter<'_>,
    ) -> Result<usize, Error> { ... }
    fn remove_xattr(&self, name: XattrName<'_>) -> Result<(), Error> { ... }
    fn check_permission(&self, perm: Permission) -> Result<(), Error> { ... }

}

Required Methods

fn size(&self) -> usize

fn resize(&self, new_size: usize) -> Result<(), Error>

fn metadata(&self) -> Metadata

fn ino(&self) -> u64

fn type_(&self) -> InodeType

fn mode(&self) -> Result<InodeMode, Error>

fn set_mode(&self, mode: InodeMode) -> Result<(), Error>

fn owner(&self) -> Result<Uid, Error>

fn set_owner(&self, uid: Uid) -> Result<(), Error>

fn group(&self) -> Result<Gid, Error>

fn set_group(&self, gid: Gid) -> Result<(), Error>

fn atime(&self) -> Duration

fn set_atime(&self, time: Duration)

fn mtime(&self) -> Duration

fn set_mtime(&self, time: Duration)

fn ctime(&self) -> Duration

fn set_ctime(&self, time: Duration)

fn fs(&self) -> Arc<dyn FileSystem>

fn extension(&self) -> &Extension

Gets the extension of this inode.

Provided Methods

fn page_cache(&self) -> Option<Arc<Vmo>>

fn create( &self, name: &str, type_: InodeType, mode: InodeMode, ) -> Result<Arc<dyn Inode>, Error>

fn mknod( &self, name: &str, mode: InodeMode, type_: MknodType, ) -> Result<Arc<dyn Inode>, Error>

fn open( &self, access_mode: AccessMode, status_flags: StatusFlags, ) -> Option<Result<Box<dyn FileIo>, Error>>

fn readdir_at( &self, offset: usize, visitor: &mut dyn DirentVisitor, ) -> Result<usize, Error>

fn link(&self, old: &Arc<dyn Inode>, name: &str) -> Result<(), Error>

fn unlink(&self, name: &str) -> Result<(), Error>

fn rmdir(&self, name: &str) -> Result<(), Error>

fn lookup(&self, name: &str) -> Result<Arc<dyn Inode>, Error>

fn rename( &self, old_name: &str, target: &Arc<dyn Inode>, new_name: &str, ) -> Result<(), Error>

fn read_link(&self) -> Result<SymbolicLink, Error>

fn write_link(&self, target: &str) -> Result<(), Error>

fn sync_all(&self) -> Result<(), Error>

fn sync_data(&self) -> Result<(), Error>

fn fallocate( &self, mode: FallocMode, offset: usize, len: usize, ) -> Result<(), Error>

Manipulates a range of space of the file according to the specified allocate mode, the manipulated range starts at offset and continues for len bytes.

fn revalidation_policy(&self) -> RevalidationPolicy

Returns the revalidation policy for cached children of this directory.

See RevalidationPolicy for the full protocol description.

Default: empty (no revalidation). Correct for filesystems where all mutations go through the VFS.

fn revalidate_exists(&self, _name: &str, _child: &dyn Inode) -> bool

Checks whether a cached child inode still exists under this directory.

Called on a positive cache hit when RevalidationPolicy::REVALIDATE_EXISTS is set. Returns true if child is still a valid child named name; false if the entry should be dropped and the lookup retried.

See RevalidationPolicy for the full protocol description.

Precondition

Only called when self.revalidation_policy() includes REVALIDATE_EXISTS. Otherwise, the return value is considered garbage.

fn revalidate_absent(&self, _name: &str) -> bool

Checks whether a name is still absent under this directory.

Called on a negative cache hit when RevalidationPolicy::REVALIDATE_ABSENT is set. Returns true if name is still absent; false if a child may now exist and the lookup should be retried.

See RevalidationPolicy for the full protocol description.

Precondition

Only called when self.revalidation_policy() includes REVALIDATE_ABSENT. Otherwise, the return value is considered garbage.

fn seek_end(&self) -> Option<usize>

Returns the end position for SeekFrom::End.

fn set_xattr( &self, name: XattrName<'_>, value_reader: &mut VmReader<'_>, flags: XattrSetFlags, ) -> Result<(), Error>

fn get_xattr( &self, name: XattrName<'_>, value_writer: &mut VmWriter<'_>, ) -> Result<usize, Error>

fn list_xattr( &self, namespace: XattrNamespace, list_writer: &mut VmWriter<'_>, ) -> Result<usize, Error>

fn remove_xattr(&self, name: XattrName<'_>) -> Result<(), Error>

fn check_permission(&self, perm: Permission) -> Result<(), Error>

Used to check for read/write/execute permissions on a file.

Similar to Linux, using “fsuid” here allows setting filesystem permissions without changing the “normal” uids for other tasks.

Implementations

impl dyn Inode

pub fn downcast_ref<T: Inode>(&self) -> Option<&T>

pub fn writer(&self, from_offset: usize) -> InodeWriter<'_>

pub fn read_bytes_at( &self, offset: usize, buf: &mut [u8], ) -> Result<usize, Error>

pub fn write_bytes_at(&self, offset: usize, buf: &[u8]) -> Result<usize, Error>