aster_kernel::fs::vfs::fs_apis::registry

Trait FsType

pub trait FsType:
    Send
    + Sync
    + 'static {
    type Key: Eq + Ord + Clone + Send + Sync + 'static;

    // Required methods
    fn name(&self) -> &'static str;
    fn properties(&self) -> FsProperties;
    fn create(
        &self,
        fs_creation_ctx: &mut FsCreationCtx<'_>,
    ) -> Result<Arc<dyn FileSystem>, Error>;
    fn sysnode(&self) -> Option<Arc<dyn SysNode>>;

    // Provided method
    fn obtain_key_and_cache(
        &self,
        _fs_creation_ctx: &mut FsCreationCtx<'_>,
    ) -> Option<(Self::Key, &FsCache<Self::Key>)> { ... }
}

A type of file system.

Required Associated Types

type Key: Eq + Ord + Clone + Send + Sync + 'static

Key used to deduplicate mounts of the same logical FS instance.

Set to () (with the default obtain_key_and_cache returning None) to opt out of any caching — every mount is fresh.

Required Methods

fn name(&self) -> &'static str

Gets the name of this FS type such as "ext4" or "sysfs".

fn properties(&self) -> FsProperties

Gets the properties of this FS type.

fn create( &self, fs_creation_ctx: &mut FsCreationCtx<'_>, ) -> Result<Arc<dyn FileSystem>, Error>

Creates an instance of this FS type.

fn sysnode(&self) -> Option<Arc<dyn SysNode>>

Returns a SysTree node that represents the FS type.

If a FS type is not intended to appear under SysFs, then this method returns a None.

The same result will be returned by this method when it is called multiple times.

Provided Methods

fn obtain_key_and_cache( &self, _fs_creation_ctx: &mut FsCreationCtx<'_>, ) -> Option<(Self::Key, &FsCache<Self::Key>)>

Returns the computed dedup key and the FsCache for this mount request.

Return None to skip the cache mechanism (every mount is fresh).

Dyn Compatibility

This trait is dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".