aster_kernel::ipc::semaphore::system_v

Module sem

Structs

AtomicStatus 🔒

PendingOp 🔒

A pending semaphore operation.

SemBuf

Semaphore

Enums

PendingBlocker 🔒

Status 🔒

Functions

check_alter_sop 🔒

Checks whether there is an operation that will change the value of a semaphore.

check_dup_sops 🔒

Checks whether there are two operations that will operate on the same semaphore and the first operation is an alteration operation.

do_smart_update 🔒

Looks for operations that can be completed after an alteration operation and then completes them.

do_smart_wakeup_zero 🔒

Wakes up pending tasks on constant operations if an alteration operation made those constant operations possible to complete.

perform_atomic_semop 🔒

Performs atomic semaphore operations.

sem_op

update_pending_alter 🔒

Looks for alteration operations that can be completed and then completes them.

wake_const_ops 🔒

Wakes up pending tasks on constant operations if they can be completed.