aster_kernel::ipc::semaphore::system_v::sem

Function check_dup_sops

fn check_dup_sops(sops: &[SemBuf]) -> bool

Checks whether there are two operations that will operate on the same semaphore and the first operation is an alteration operation.