aster_kernel::ipc::semaphore::system_v

Module sem_set

Structs

IpcPerm

SemSetInner 🔒

SemaphoreSet

SemidDs

Constants

SEMAEM

Maximum value that can be recorded for semaphore adjustment (SEM_UNDO).

SEMMNI

Maximum number of semaphore sets.

SEMMNS

Maximum number of seaphores in all semaphore sets.

SEMMSL

Maximum number of semaphores per semaphore ID.

SEMOPM

Maximum number of operations for semop.

SEMVMX

MAximum semaphore value.

Statics

ID_ALLOCATOR 🔒

SEMAPHORE_SETS 🔒

Semaphore sets in system

Functions

check_sem

Checks the semaphore. Return Ok if the semaphore exists and pass the check.

create_sem_set

create_sem_set_with_id

init_in_first_kthread 🔒

sem_sets

sem_sets_mut