aster_kernel::net::socket::ip::common

Function get_ephemeral_iface

fn get_ephemeral_iface(remote_ip_addr: &IpAddress) -> Arc<dyn Iface<BigtcpExt>> 

Get a suitable iface to deal with sendto/connect request if the socket is not bound to an iface. If the remote address is the same as that of some iface, we will use the iface. Otherwise, we will use a default interface.