aster_kernel::net::socket::ip::stream

Struct StreamSocket

pub struct StreamSocket {
    state: RwLock<Takeable<State>>,
    options: RwLock<OptionSet>,
    is_nonblocking: AtomicBool,
    pollee: Pollee,
    pseudo_path: Path,
}

Fields
state: RwLock<Takeable<State>>options: RwLock<OptionSet>is_nonblocking: AtomicBoolpollee: Polleepseudo_path: Path
Implementations

impl StreamSocket

pub fn new(is_nonblocking: bool) -> Arc<Self>

fn new_accepted(connected_stream: ConnectedStream) -> Arc<Self>

fn read_updated_state( &self, ) -> RwLockReadGuard<'_, Takeable<State>, PreemptDisabled>

Ensures that the socket state is up to date and obtains a read lock on it.

For a description of what “up-to-date” means, see Self::write_updated_state.

fn write_updated_state( &self, ) -> RwLockWriteGuard<'_, Takeable<State>, PreemptDisabled>

Ensures that the socket state is up to date and obtains a write lock on it.

A connecting socket can become obsolete because some network events can set the socket to connected state (if the connection succeeds) or initial state (if the connection is refused) in Self::check_io_events, but the state transition is delayed until the user operates on the socket to avoid too many locks in the interrupt handler.

This method performs the delayed state transition to ensure that the state is up to date and returns the guard of the write-locked state.

fn start_connect( &self, remote_endpoint: &IpEndpoint, ) -> Option<Result<(), Error>>

fn check_connect(&self) -> Result<(), Error>

fn try_accept(&self) -> Result<(Arc<dyn FileLike>, SocketAddr), Error>

fn try_recv( &self, writer: &mut dyn MultiWrite, flags: SendRecvFlags, ) -> Result<(usize, SocketAddr), Error>

fn try_send( &self, reader: &mut dyn MultiRead, flags: SendRecvFlags, ) -> Result<usize, Error>

fn check_io_events(&self) -> IoEvents

fn test_and_clear_error(&self) -> Option<Error>