aster_kernel::net::socket::netlink::message::attr

Trait Attribute

pub trait Attribute:
    Debug
    + Send
    + Sync {
    // Required methods
    fn type_(&self) -> u16;
    fn payload_as_bytes(&self) -> &[u8];
    fn read_from(
        header: &CAttrHeader,
        reader: &mut dyn MultiRead,
    ) -> Result<ContinueRead<Self>, Error>
       where Self: Sized;

    // Provided methods
    fn total_len_with_padding(&self) -> usize { ... }
    fn read_all_from(
        reader: &mut dyn MultiRead,
        total_len: usize,
    ) -> Result<ContinueRead<Vec<Self>>, Error>
       where Self: Sized { ... }
    fn write_to(&self, writer: &mut dyn MultiWrite) -> Result<(), Error> { ... }
}

Netlink Attribute.

Required Methods

fn type_(&self) -> u16

Returns the type of the attribute.

fn payload_as_bytes(&self) -> &[u8]

Returns the byte representation of the payload.

fn read_from( header: &CAttrHeader, reader: &mut dyn MultiRead, ) -> Result<ContinueRead<Self>, Error>
where Self: Sized,

Reads the attribute from the reader.

This method may return a ContinueRead::Skipped or a ContinueRead::SkippedErr if the attribute is not recognized or not valid. In that case, however, it must still skip the payload length (excluding padding), as if the attribute were parsed properly.

Provided Methods

fn total_len_with_padding(&self) -> usize

Returns the total length of the attribute (header + payload, including padding).

fn read_all_from( reader: &mut dyn MultiRead, total_len: usize, ) -> Result<ContinueRead<Vec<Self>>, Error>
where Self: Sized,

Reads all attributes from the reader.

The cumulative length of the read attributes must not exceed total_len.

fn write_to(&self, writer: &mut dyn MultiWrite) -> Result<(), Error>

Writes the attribute to the writer.