aster_kernel::net::socket::netlink::message

Trait ProtocolSegment

pub trait ProtocolSegment: Sized {
    // Required methods
    fn header(&self) -> &CMsgSegHdr;
    fn header_mut(&mut self) -> &mut CMsgSegHdr;
    fn read_from(
        reader: &mut dyn MultiRead,
    ) -> Result<ContinueRead<Self, SegmentCommon<ErrorSegmentBody, NoAttr>>, Error>;
    fn write_to(&self, writer: &mut dyn MultiWrite) -> Result<(), Error>;
}

Required Methods

fn header(&self) -> &CMsgSegHdr

fn header_mut(&mut self) -> &mut CMsgSegHdr

fn read_from( reader: &mut dyn MultiRead, ) -> Result<ContinueRead<Self, SegmentCommon<ErrorSegmentBody, NoAttr>>, Error>

Reads the segment body from the reader.

If the reader encounters an unresolvable page fault, this method will fail with Errno::EFAULT. Netlink sockets should directly report this error code to the user.

If the reader does not contain a valid segment header (CMsgSegHdr), this method will also fail. Netlink sockets should then stop parsing segments from the reader and silently ignore the error.

If the reader contains a valid segment header but an invalid segment (e.g., one with an invalid body or attributes), this method will succeed with ContinueRead::Skipped or ContinueRead::SkippedErr. If there is an error segment in ContinueRead::SkippedErr, netlink sockets should respond the user with the error segment. The entire segment is skipped so it is possible to read the next segment from the reader.

This method will skip the padding bytes, so it can be called multiple times to read multiple segments.

fn write_to(&self, writer: &mut dyn MultiWrite) -> Result<(), Error>

Writes the segment to the writer.

This method will skip the padding bytes, so it can be called multiple times to write multiple segments.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.