aster_kernel::net::socket::options

Trait SocketOption

pub trait SocketOption:
    Any
    + Send
    + Sync
    + Debug {
    // Required methods
    fn as_any(&self) -> &dyn Any;
    fn as_any_mut(&mut self) -> &mut dyn Any;
}

Socket options. This trait represents all options that can be set or got for a socket, including socket level options and options for specific socket type like tcp socket.

Required Methods

fn as_any(&self) -> &dyn Any

fn as_any_mut(&mut self) -> &mut dyn Any