aster_kernel::net::socket::private

Trait SocketPrivate

pub trait SocketPrivate: Pollable {
    // Required methods
    fn is_nonblocking(&self) -> bool;
    fn set_nonblocking(&self, nonblocking: bool);

    // Provided method
    fn block_on<F, R>(&self, events: IoEvents, try_op: F) -> Result<R, Error>
       where Self: Sized,
             F: FnMut() -> Result<R, Error> { ... }
}

Common methods for sockets, but private to the network module.

These are implementation details of sockets, so shouldn’t be accessed outside the network module. Therefore, the whole trait is sealed.

Required Methods

fn is_nonblocking(&self) -> bool

Returns whether the socket is in non-blocking mode.

fn set_nonblocking(&self, nonblocking: bool)

Sets whether the socket is in non-blocking mode.

Provided Methods

fn block_on<F, R>(&self, events: IoEvents, try_op: F) -> Result<R, Error>
where Self: Sized, F: FnMut() -> Result<R, Error>,

Blocks until some events occur to complete I/O operations.

If the socket is in non-blocking mode and the I/O operations cannot be completed immediately, this method will fail with EAGAIN instead of blocking.