aster_kernel::net::socket

Trait Socket

pub trait Socket:
    SocketPrivate
    + Send
    + Sync {
    // Required methods
    fn sendmsg(
        &self,
        reader: &mut dyn MultiRead,
        message_header: MessageHeader,
        flags: SendRecvFlags,
    ) -> Result<usize, Error>;
    fn recvmsg(
        &self,
        writer: &mut dyn MultiWrite,
        flags: SendRecvFlags,
    ) -> Result<(usize, MessageHeader), Error>;
    fn pseudo_path(&self) -> &Path;

    // Provided methods
    fn bind(&self, _socket_addr: SocketAddr) -> Result<(), Error> { ... }
    fn connect(&self, _socket_addr: SocketAddr) -> Result<(), Error> { ... }
    fn listen(&self, _backlog: usize) -> Result<(), Error> { ... }
    fn accept(&self) -> Result<(Arc<dyn FileLike>, SocketAddr), Error> { ... }
    fn shutdown(&self, _cmd: SockShutdownCmd) -> Result<(), Error> { ... }
    fn addr(&self) -> Result<SocketAddr, Error> { ... }
    fn peer_addr(&self) -> Result<SocketAddr, Error> { ... }
    fn get_option(&self, _option: &mut dyn SocketOption) -> Result<(), Error> { ... }
    fn set_option(&self, _option: &dyn SocketOption) -> Result<(), Error> { ... }
}

Operations defined on a socket.

Required Methods

fn sendmsg( &self, reader: &mut dyn MultiRead, message_header: MessageHeader, flags: SendRecvFlags, ) -> Result<usize, Error>

Sends a message on the socket.

fn recvmsg( &self, writer: &mut dyn MultiWrite, flags: SendRecvFlags, ) -> Result<(usize, MessageHeader), Error>

Receives a message from the socket.

If successful, the io_vecs buffer will be filled with the received content. This method returns the length of the received message, and the message header.

fn pseudo_path(&self) -> &Path

Returns a reference to the pseudo path associated with this socket.

Provided Methods

fn bind(&self, _socket_addr: SocketAddr) -> Result<(), Error>

Assigns the specified address to the socket.

fn connect(&self, _socket_addr: SocketAddr) -> Result<(), Error>

Builds a connection for the given address

fn listen(&self, _backlog: usize) -> Result<(), Error>

Listens for connections on the socket.

fn accept(&self) -> Result<(Arc<dyn FileLike>, SocketAddr), Error>

Accepts a connection on the socket.

fn shutdown(&self, _cmd: SockShutdownCmd) -> Result<(), Error>

Shuts down part of a full-duplex connection.

fn addr(&self) -> Result<SocketAddr, Error>

Gets the address of this socket.

fn peer_addr(&self) -> Result<SocketAddr, Error>

Gets the address of the peer socket.

fn get_option(&self, _option: &mut dyn SocketOption) -> Result<(), Error>

Gets options on the socket.

If the method succeeds, the result will be stored in the option parameter.

fn set_option(&self, _option: &dyn SocketOption) -> Result<(), Error>

Sets options on the socket.