aster_kernel::net::socket::util::options

Trait SetSocketLevelOption

pub(in net) trait SetSocketLevelOption {
    // Provided methods
    fn set_reuse_addr(&self, _reuse_addr: bool) { ... }
    fn set_keep_alive(&self, _keep_alive: bool) -> NeedIfacePoll { ... }
    fn set_pass_cred(&self, _pass_cred: bool) { ... }
}

A trait used for setting socket level options on actual sockets.

Provided Methods

fn set_reuse_addr(&self, _reuse_addr: bool)

Sets whether the socket address can be reused.

fn set_keep_alive(&self, _keep_alive: bool) -> NeedIfacePoll

Sets whether keepalive messages are enabled.

fn set_pass_cred(&self, _pass_cred: bool)

Sets whether receipt of the credentials of the sending process is enabled.