aster_kernel::net::socket

Module vsock

This module defines vsock sockets.

The vsock address family facilitates communication between virtual machines and the host they are running on. This address family is used by guest agents and hypervisor services that need a communications channel that is independent of virtual machine network configuration.

The implementation is organized into three layers:

The device layer provides the basic packet transmit and receive primitives.

The transport layer implements protocol logic such as connection and listener management.

The socket layer builds the Linux-compatible socket interface used by userspace-facing system calls.

Re-exports

pub use addr::VsockSocketAddr;

pub use stream::VsockStreamSocket;

Modules

addr 🔒

stream 🔒

transport 🔒

The virtio-vsock transport protocol.

Functions

init 🔒