aster_kernel::net::socket::vsock

Module transport

The virtio-vsock transport protocol.

Built on top of the basic virtio-vsock device support, this module manages connections and listeners, handles connection establishment, data transfer, and shutdown, and implements credit-based flow control plus I/O event checking and notification. The socket layer is expected to build on these APIs to provide the user-visible socket interface.

For a quick start, bind to a port by creating a BoundPort instance.

To connect to a remote address, use the BoundPort::connect method and get a Connection instance. Data can be transmitted or received via the Connection::try_send and Connection::try_recv methods.

To listen to the local address, use the BoundPort::listen method and get a Listener instance. Incoming connections can be accepted via the Listener::try_accept method.

Drop the BoundPort, Connection, or Listener instance will shut down the underlying connection (if any) and release the resources.

Modules

conn_id 🔒

connection 🔒

listener 🔒

port 🔒

space 🔒

timer 🔒

Constants

CREDIT_UPDATE_THRESHOLD 🔒

DEFAULT_CLOSE_TIMEOUT 🔒

DEFAULT_CONNECT_TIMEOUT 🔒

DEFAULT_RX_BUF_SIZE 🔒

DEFAULT_TX_BUF_SIZE 🔒

MAX_BACKLOG 🔒

Functions

init 🔒

Initializes the virtio-vsock transport when the default device is present.

process_event_callback 🔒

process_rx_callback 🔒