aster_kernel

Module prelude

Re-exports

pub use crate::context::Context;

pub use crate::error::Errno;

pub use crate::error::Error;

pub use crate::process::posix_thread::AsThreadLocal;

pub use crate::process::signal::Pause;

pub use crate::time::Clock;

pub use crate::time::wait::WaitTimeout;

pub use crate::util::ReadCString;

Modules

vec

A contiguous growable array type with heap-allocated contents, written Vec<T>.

Macros

alert

Logs a message at the Alert level.

bitflags

The macro used to generate the flag structures.

crit

Logs a message at the Crit level.

debug

Logs a message at the Debug level.

emerg

Logs a message at the Emerg level.

error

Logs a message at the Error level.

info

Logs a message at the Info level.

notice

Logs a message at the Notice level.

print

Copied from Rust std: https://github.com/rust-lang/rust/blob/master/library/std/src/macros.rs

println

Copied from Rust std: https://github.com/rust-lang/rust/blob/master/library/std/src/macros.rs

vec

Creates a Vec containing the arguments.

warn

Logs a message at the Warning level.

Structs

Arc

A thread-safe reference-counting pointer. ‘Arc’ stands for ‘Atomically Reference Counted’.

BTreeMap

An ordered map based on a B-Tree.

BTreeSet

An ordered set based on a B-Tree.

Box

A pointer type that uniquely owns a heap allocation of type T.

CStr

A dynamically-sized view of a C string.

CString

A type representing an owned, C-compatible, nul-terminated string with no nul bytes in the middle.

LinkedList

A doubly-linked list with owned nodes.

Mutex

A mutex with waitqueue.

MutexGuard

A guard that provides exclusive access to the data protected by a Mutex.

RwLock

Spin-based Read-write Lock

RwMutex

A mutex that provides data access to either one writer or many readers.

SpinLock

A spin lock.

SpinLockGuard

A guard that provides exclusive access to the data protected by a SpinLock.

String

A UTF-8–encoded, growable string.

Vec

A contiguous growable array type, written as Vec<T>, short for ‘vector’.

VecDeque

A double-ended queue implemented with a growable ring buffer.

VmReader

VmReader is a reader for reading data from a contiguous range of memory.

VmWriter

VmWriter is a writer for writing data to a contiguous range of memory.

Weak

Weak is a version of Arc that holds a non-owning reference to the managed allocation.

Constants

PAGE_SIZE

The page size

Traits

Any

A trait to emulate dynamic typing.

Debug

? formatting.

FallibleVmRead

Fallible memory read from a VmWriter.

FallibleVmWrite

Fallible memory write from a VmReader.

FromBytes

Types for which any bit pattern is valid.

FromZeros

Types for which a sequence of 0 bytes is a valid instance.

IntoBytes

Types that can be converted to an immutable slice of initialized bytes.

Pod

A trait for plain old data (POD).

ToString

A trait for converting a value to a String.

Type Aliases

Result

Vaddr

Virtual addresses.

Derive Macros

Debug

Derive macro generating an impl of the trait Debug.

FromBytes

FromZeros

IntoBytes

TryFromInt