aster_kernel

Module process

Re-exports

pub use clone::CloneArgs;

pub use clone::CloneFlags;

pub use clone::clone_child;

pub use credentials::Credentials;

pub use credentials::Gid;

pub use credentials::Uid;

pub use execve::do_execve;

pub use kill::kill;

pub use kill::kill_all;

pub use kill::kill_group;

pub use kill::tgkill;

pub use namespace::nsproxy::ContextSetNsAdminApi;

pub use namespace::nsproxy::NsProxy;

pub use namespace::nsproxy::NsProxyBuilder;

pub use namespace::nsproxy::check_unsupported_ns_flags;

pub use namespace::unshare::ContextUnshareAdminApi;

pub use namespace::user_ns::UserNamespace;

pub use pid_file::PidFile;

pub use process::ExitCode;

pub use process::JobControl;

pub use process::Pgid;

pub use process::Pid;

pub use process::Process;

pub use process::ProcessGroup;

pub use process::ReapedChildrenStats;

pub use process::Session;

pub use process::Sid;

pub use process::Terminal;

pub use process::broadcast_signal_async;

pub use process::enqueue_signal_async;

pub use process::spawn_init_process;

pub use process_filter::ProcessFilter;

pub use process_vm::INIT_STACK_SIZE;

pub use process_vm::LockedHeap;

pub use process_vm::ProcessVm;

pub use rlimit::ResourceType;

pub use stats::collect_process_creation_count;

pub use term_status::TermStatus;

pub use wait::WaitOptions;

pub use wait::WaitStatus;

pub use wait::do_wait;

Modules

clone 🔒

credentials

execve 🔒

exit 🔒

kill 🔒

namespace 🔒

pid_file 🔒

posix_thread

process 🔒

process_filter 🔒

process_table

A global table stores the pid to process mapping. This table can be used to get process with pid. TODO: progress group, thread all need similar mapping

process_vm 🔒

This module defines struct ProcessVm to represent the layout of user space process virtual memory.

program_loader 🔒

rlimit

signal

stats 🔒

status 🔒

The process status.

sync

task_set

Task sets.

term_status 🔒

wait 🔒

Functions

init 🔒

init_in_first_process 🔒

init_on_each_cpu 🔒