aster_kernel::process::credentials::credentials_

Struct Credentials_

pub(super) struct Credentials_ {

    ruid: AtomicUid,
    euid: AtomicUid,
    suid: AtomicUid,
    fsuid: AtomicUid,
    rgid: AtomicGid,
    egid: AtomicGid,
    sgid: AtomicGid,
    fsgid: AtomicGid,
    supplementary_gids: RwLock<BTreeSet<Gid>>,
    inheritable_capset: AtomicCapSet,
    permitted_capset: AtomicCapSet,
    effective_capset: AtomicCapSet,
    securebits: AtomicSecureBits,

}

Fields
ruid: AtomicUid

The real user ID.

This is the user to which the process belongs.

euid: AtomicUid

The effective user ID.

This is used to determine the permissions granted to a process when it tries to perform various operations (e.g., system calls).

suid: AtomicUid

The saved-set user ID.

This saves a copy of the effective user ID that were set when the program was executed.

fsuid: AtomicUid

The filesystem user ID.

This is used to determine permissions for accessing files.

rgid: AtomicGid

The real group ID.

This is the group to which the process belongs.

egid: AtomicGid

The effective group ID.

This is used to determine the permissions granted to a process when it tries to perform various operations (e.g., system calls).

sgid: AtomicGid

The saved-set group ID.

This saves a copy of the effective group ID that were set when the program was executed.

fsgid: AtomicGid

The filesystem group ID.

This is used to determine permissions for accessing files.

supplementary_gids: RwLock<BTreeSet<Gid>>

A set of additional groups to which a process belongs.

inheritable_capset: AtomicCapSet

Capabilities that child processes can inherit.

permitted_capset: AtomicCapSet

Capabilities that a process can potentially be granted.

It defines the maximum set of privileges that the process could possibly have. Even if the process is not currently using these privileges, it has the potential ability to enable them.

effective_capset: AtomicCapSet

Capabilities that we can actually use.

securebits: AtomicSecureBits

Secure bits.

Implementations

impl Credentials_

pub(super) fn new(uid: Uid, gid: Gid, capset: CapSet) -> Self

Creates a new Credentials_.

The real, effective, saved set, and filesystem IDs will be initialized to the same ID.

pub(super) fn ruid(&self) -> Uid

pub(super) fn euid(&self) -> Uid

pub(super) fn suid(&self) -> Uid

pub(super) fn fsuid(&self) -> Uid

pub(super) fn set_uid(&self, uid: Uid) -> Result<(), Error>

pub(super) fn set_reuid( &self, ruid: Option<Uid>, euid: Option<Uid>, ) -> Result<(), Error>

pub(super) fn set_resuid( &self, ruid: Option<Uid>, euid: Option<Uid>, suid: Option<Uid>, ) -> Result<(), Error>

pub(super) fn set_fsuid(&self, fsuid: Option<Uid>) -> Result<Uid, Uid>

pub(super) fn set_euid(&self, euid: Uid)

pub(super) fn set_suid(&self, suid: Uid)

fn check_uid_perm( &self, ruid: Option<&Uid>, euid: Option<&Uid>, suid: Option<&Uid>, ruid_may_be_old_suid: bool, ) -> Result<(), Error>

fn set_resuid_unchecked( &self, ruid: Option<Uid>, euid: Option<Uid>, suid: Option<Uid>, )

fn set_fsuid_unchecked(&self, fsuid: Uid)

pub(super) fn rgid(&self) -> Gid

pub(super) fn egid(&self) -> Gid

pub(super) fn sgid(&self) -> Gid

pub(super) fn fsgid(&self) -> Gid

pub(super) fn set_gid(&self, gid: Gid) -> Result<(), Error>

pub(super) fn set_regid( &self, rgid: Option<Gid>, egid: Option<Gid>, ) -> Result<(), Error>

pub(super) fn set_resgid( &self, rgid: Option<Gid>, egid: Option<Gid>, sgid: Option<Gid>, ) -> Result<(), Error>

pub(super) fn set_fsgid(&self, fsgid: Option<Gid>) -> Result<Gid, Gid>

pub(super) fn set_egid(&self, egid: Gid)

pub(super) fn set_sgid(&self, sgid: Gid)

fn check_gid_perm( &self, rgid: Option<&Gid>, egid: Option<&Gid>, sgid: Option<&Gid>, rgid_may_be_old_sgid: bool, ) -> Result<(), Error>

fn set_resgid_unchecked( &self, rgid: Option<Gid>, egid: Option<Gid>, sgid: Option<Gid>, )

fn set_fsgid_unchecked(&self, fsuid: Gid)

pub(super) fn groups( &self, ) -> RwLockReadGuard<'_, BTreeSet<Gid>, PreemptDisabled>

pub(super) fn groups_mut( &self, ) -> RwLockWriteGuard<'_, BTreeSet<Gid>, PreemptDisabled>

pub(super) fn inheritable_capset(&self) -> CapSet

pub(super) fn permitted_capset(&self) -> CapSet

pub(super) fn effective_capset(&self) -> CapSet

pub(super) fn set_inheritable_capset(&self, inheritable_capset: CapSet)

pub(super) fn set_permitted_capset(&self, permitted_capset: CapSet)

pub(super) fn set_effective_capset(&self, effective_capset: CapSet)

pub(super) fn keep_capabilities(&self) -> bool

pub(super) fn set_keep_capabilities( &self, keep_capabilities: bool, ) -> Result<(), Error>

pub(super) fn securebits(&self) -> SecureBits

pub(super) fn set_securebits(&self, securebits: SecureBits) -> Result<(), Error>