aster_kernel::process::credentials

Struct Credentials

pub struct Credentials<R = FullOp>(Arc<Credentials_>, R);

A set of associated numeric user IDs (UIDs) and group IDs (GIDs) for a process.

This type contains:

real user ID and group ID;

effective user ID and group ID;

saved-set user ID and saved-set group ID;

filesystem user ID and group ID (Linux-specific);

supplementary group IDs;

Linux capabilities;

secure bits.

Tuple Fields
0: Arc<Credentials_>1: R
Implementations

impl<R: TRights> Credentials<R>

pub fn new_root() -> Self

Creates a root Credentials.

This method can only be used when creating the init process.

pub fn new_from<R1>(credentials: &Credentials<R1>) -> Self
where R1: SetContain<Read> + TRights, SetContainOp<R1, Read>: IsTrue,

Clones a new Credentials from an existing Credentials.

This method requires the Read right.

pub fn dup(&self) -> Self
where R: SetContain<Dup>, SetContainOp<R, Dup>: IsTrue,

Duplicates the capabilities.

This method requires the Dup right.

pub fn restrict<R1>(self) -> Credentials<R1>
where R1: Set + TRights, R: SetInclude<R1>, SetIncludeOp<R, R1>: IsTrue,

Restricts capabilities to a smaller set.

pub fn ruid(&self) -> Uid
where R: SetContain<Read>, SetContainOp<R, Read>: IsTrue,

Gets the real user ID.

This method requires the Read right.

pub fn euid(&self) -> Uid
where R: SetContain<Read>, SetContainOp<R, Read>: IsTrue,

Gets the effective user ID.

This method requires the Read right.

pub fn suid(&self) -> Uid
where R: SetContain<Read>, SetContainOp<R, Read>: IsTrue,

Gets the saved-set user ID.

This method requires the Read right.

pub fn fsuid(&self) -> Uid
where R: SetContain<Read>, SetContainOp<R, Read>: IsTrue,

Gets the filesystem user ID.

This method requires the Read right.

pub fn set_uid(&self, uid: Uid) -> Result<(), Error>
where R: SetContain<Write>, SetContainOp<R, Write>: IsTrue,

Sets the user ID.

If self is privileged, sets the real, effective, saved-set user IDs as uid, Otherwise, sets the effective user ID as uid.

This method requires the Write right.

pub fn set_reuid( &self, ruid: Option<Uid>, euid: Option<Uid>, ) -> Result<(), Error>
where R: SetContain<Write>, SetContainOp<R, Write>: IsTrue,

Sets the real, effective user IDs as ruid, euid respectively.

If ruid or euid is None, the corresponding user ID will leave unchanged.

This method requires the Write right.

pub fn set_resuid( &self, ruid: Option<Uid>, euid: Option<Uid>, suid: Option<Uid>, ) -> Result<(), Error>
where R: SetContain<Write>, SetContainOp<R, Write>: IsTrue,

Sets the real, effective, saved-set user IDs as ruid, euid, suid respectively.

If ruid, euid, or suid is None, the corresponding user ID will leave unchanged.

This method requires the Write right.

pub fn set_fsuid(&self, fsuid: Option<Uid>) -> Result<Uid, Uid>
where R: SetContain<Write>, SetContainOp<R, Write>: IsTrue,

Sets the filesystem user ID as fsuid and returns the original filesystem user ID.

If fsuid is None, leaves the filesystem user ID unchanged.

This method requires the Write right.

pub fn set_euid(&self, euid: Uid)
where R: SetContain<Write>, SetContainOp<R, Write>: IsTrue,

Sets the effective user ID as euid.

This method should only be used when executing a file whose setuid bit is set.

This method requires the Write right.

pub fn reset_suid(&self)
where R: SetContain<Write>, SetContainOp<R, Write>: IsTrue,

Sets the saved-set user ID as the same of the effective user ID.

This method should only be used when executing a new executable file.

This method requires the Write right.

pub fn rgid(&self) -> Gid
where R: SetContain<Read>, SetContainOp<R, Read>: IsTrue,

Gets the real group ID.

This method requires the Read right.

pub fn egid(&self) -> Gid
where R: SetContain<Read>, SetContainOp<R, Read>: IsTrue,

Gets the effective group ID.

This method requires the Read right.

pub fn sgid(&self) -> Gid
where R: SetContain<Read>, SetContainOp<R, Read>: IsTrue,

Gets the saved-set group ID.

This method requires the Read right.

pub fn fsgid(&self) -> Gid
where R: SetContain<Read>, SetContainOp<R, Read>: IsTrue,

Gets the filesystem group ID.

This method requires the Read right.

pub fn set_gid(&self, gid: Gid) -> Result<(), Error>
where R: SetContain<Write>, SetContainOp<R, Write>: IsTrue,

Sets the group ID.

If self is privileged, sets the real, effective, saved-set group IDs as gid, Otherwise, sets the effective group ID as gid.

This method requires the Write right.

pub fn set_regid( &self, rgid: Option<Gid>, egid: Option<Gid>, ) -> Result<(), Error>
where R: SetContain<Write>, SetContainOp<R, Write>: IsTrue,

Sets the real, effective group IDs as rgid, egid respectively.

If rgid or egid is None, the corresponding group ID will leave unchanged.

This method requires the Write right.

pub fn set_resgid( &self, rgid: Option<Gid>, egid: Option<Gid>, sgid: Option<Gid>, ) -> Result<(), Error>
where R: SetContain<Write>, SetContainOp<R, Write>: IsTrue,

Sets the real, effective, saved-set group IDs as rgid, egid, sgid respectively.

If rgid, egid, or sgid is None, the corresponding group ID will leave unchanged.

This method requires the Write right.

pub fn set_fsgid(&self, fsgid: Option<Gid>) -> Result<Gid, Gid>
where R: SetContain<Write>, SetContainOp<R, Write>: IsTrue,

Sets the filesystem group ID as fsgid and returns the original filesystem group ID.

If fsgid is None, leaves the filesystem group ID unchanged.

This method requires the Write right.

pub fn set_egid(&self, egid: Gid)
where R: SetContain<Write>, SetContainOp<R, Write>: IsTrue,

Sets the effective group ID as egid.

This method should only be used when executing a file whose setgid bit is set.

This method requires the Write right.

pub fn reset_sgid(&self)
where R: SetContain<Write>, SetContainOp<R, Write>: IsTrue,

Sets the saved-set group ID as the same of the effective group ID.

This method should only be used when executing a new executable file.

This method requires the Write right.

pub fn groups(&self) -> RwLockReadGuard<'_, BTreeSet<Gid>, PreemptDisabled>
where R: SetContain<Read>, SetContainOp<R, Read>: IsTrue,

Acquires the read lock of supplementary group IDs.

This method requires the Read right.

pub fn groups_mut(&self) -> RwLockWriteGuard<'_, BTreeSet<Gid>, PreemptDisabled>
where R: SetContain<Write>, SetContainOp<R, Write>: IsTrue,

Acquires the write lock of supplementary group IDs.

This method requires the Write right.

pub fn inheritable_capset(&self) -> CapSet
where R: SetContain<Read>, SetContainOp<R, Read>: IsTrue,

Gets the capabilities that child processes can inherit.

This method requires the Read right.

pub fn permitted_capset(&self) -> CapSet
where R: SetContain<Read>, SetContainOp<R, Read>: IsTrue,

Gets the capabilities that are a process can potentially be granted.

This method requires the Read right.

pub fn effective_capset(&self) -> CapSet
where R: SetContain<Read>, SetContainOp<R, Read>: IsTrue,

Gets the capabilities that we can actually use.

This method requires the Read right.

pub fn set_inheritable_capset(&self, inheritable_capset: CapSet)
where R: SetContain<Write>, SetContainOp<R, Write>: IsTrue,

Sets the capabilities that child processes can inherit.

This method requires the Write right.

pub fn set_permitted_capset(&self, permitted_capset: CapSet)
where R: SetContain<Write>, SetContainOp<R, Write>: IsTrue,

Sets the capabilities that are a process can potentially be granted.

This method requires the Write right.

pub fn set_effective_capset(&self, effective_capset: CapSet)
where R: SetContain<Write>, SetContainOp<R, Write>: IsTrue,

Sets the capabilities that we can actually use.

This method requires the Write right.

pub fn keep_capabilities(&self) -> bool
where R: SetContain<Read>, SetContainOp<R, Read>: IsTrue,

Gets keep capabilities flag.

This method requires the Read right.

pub fn set_keep_capabilities( &self, keep_capabilities: bool, ) -> Result<(), Error>
where R: SetContain<Write>, SetContainOp<R, Write>: IsTrue,

Sets keep capabilities flag.

If the SecureBits::KEEP_CAPS_LOCKED is set, this method will return an error.

This method requires the Write right.

pub fn securebits(&self) -> SecureBits
where R: SetContain<Read>, SetContainOp<R, Read>: IsTrue,

Gets the secure bits.

This method requires the Read right.

pub fn set_securebits(&self, securebits: SecureBits) -> Result<(), Error>
where R: SetContain<Write>, SetContainOp<R, Write>: IsTrue,

Sets the secure bits.

If the caller does not have the CAP_SETPCAP capability, or if it tries to set locked bits, this method will return an error.

This method requires the Write right.