aster_kernel::process::exit

Macro drop_after

macro_rules! drop_after {
    ($expr:expr) => { ... };
}

Drops a value after releasing the lock or borrow guard.

For example, ThreadLocal::vmar needs to be borrowed during a context switch. This means that the kernel may panic if we borrow it as mutable and tries to acquire a mutex while dropping the last VmarHandle. This utility macro ensures that the value is dropped only after the borrow guard has been released, preventing all these issues.