aster_kernel::process

Module kill

Functions

check_signal_perm 🔒

kill

Sends a signal to a process, using the current process as the sender.

kill_all

Sends a signal to all processes except current process and init process, using the current process as the sender.

kill_group

Sends a signal to all processes in a group, using the current process as the sender.

kill_process 🔒

tgkill

Sends a signal to a target thread, using the current process as the sender.