aster_kernel::process::posix_thread

Struct PosixThread

pub struct PosixThread {

    process: Weak<Process>,
    task: Weak<Task>,
    tid: AtomicU32,
    name: Mutex<ThreadName>,
    credentials: Credentials,
    fs: RwMutex<Arc<ThreadFsInfo>>,
    file_table: Mutex<Option<RoArc<FileTable>>>,
    sig_mask: AtomicSigSet,
    sig_queues: SigQueues,
    signalled_waker: SpinLock<Option<(Arc<Waker>, PauseReason)>>,
    prof_clock: Arc<ProfClock>,
    virtual_timer_manager: Arc<TimerManager>,
    prof_timer_manager: Arc<TimerManager>,
    io_priority: AtomicU32,
    ns_proxy: Mutex<Option<Arc<NsProxy>>>,
    timer_slack_ns: AtomicU64,
    default_timer_slack_ns: AtomicU64,

}

Fields
process: Weak<Process>task: Weak<Task>tid: AtomicU32name: Mutex<ThreadName>credentials: Credentials

Process credentials. At the kernel level, credentials are a per-thread attribute.

fs: RwMutex<Arc<ThreadFsInfo>>

The file system information of the thread.

file_table: Mutex<Option<RoArc<FileTable>>>

File table

sig_mask: AtomicSigSet

Blocked signals

sig_queues: SigQueues

Thread-directed sigqueue

signalled_waker: SpinLock<Option<(Arc<Waker>, PauseReason)>>

The per-thread signal Waker, which will be used to wake up the thread when enqueuing a signal, along with the reason why the thread is paused.

prof_clock: Arc<ProfClock>

A profiling clock measures the user CPU time and kernel CPU time in the thread.

virtual_timer_manager: Arc<TimerManager>

A manager that manages timers based on the user CPU time of the current thread.

prof_timer_manager: Arc<TimerManager>

A manager that manages timers based on the profiling clock of the current thread.

io_priority: AtomicU32

I/O Scheduling priority value

ns_proxy: Mutex<Option<Arc<NsProxy>>>

The namespaces that the thread belongs to.

timer_slack_ns: AtomicU64

The current timer slack value for this thread.

default_timer_slack_ns: AtomicU64

The default timer slack value for this thread.

Implementations

impl PosixThread

pub fn process(&self) -> Arc<Process>

pub fn weak_process(&self) -> Weak<Process>

pub fn tid(&self) -> u32

Returns the thread id

pub(super) fn set_main(&self, pid: u32)

Sets the thread as the main thread by changing its thread ID.

pub fn thread_name(&self) -> &Mutex<ThreadName>

pub fn read_fs(&self) -> RwMutexReadGuard<'_, Arc<ThreadFsInfo>>

Returns a read guard to the filesystem information of the thread.

pub(super) fn set_fs(&self, new_fs: Arc<ThreadFsInfo>)

Sets the filesystem information of the thread.

pub fn file_table(&self) -> &Mutex<Option<RoArc<FileTable>>>

pub fn sig_mask(&self) -> &AtomicSigSet

Gets the reference to the signal mask of the thread.

Note that while this function offers mutable access to the signal mask, it is not sound for callers other than the current thread to modify the signal mask. They may only read the signal mask.

pub(super) fn sig_queues(&self) -> &SigQueues

pub fn has_signal_blocked(&self, signum: SigNum) -> bool

Returns whether the signal is blocked by the thread.

pub fn set_signalled_waker(&self, waker: Arc<Waker>, reason: PauseReason)

Sets the input Waker as the signalled waker of this thread, along with the reason why the thread is paused.

This approach can collaborate with signal-aware wait methods. Once a signalled waker is set for a thread, it cannot be reset until it is cleared.

Panics

If setting a new waker before clearing the current thread’s signalled waker this method will panic.

pub fn clear_signalled_waker(&self)

Clears the signalled waker of this thread.

pub fn sleeping_state(&self) -> SleepingState

Returns the sleeping state of this thread.

pub fn wake_signalled_waker(&self)

Wakes up the signalled waker.

pub fn enqueue_signal(&self, signal: Box<dyn Signal>)

Enqueues a thread-directed signal.

This method does not perform permission checks on user signals. Therefore, unless the caller can ensure that there are no permission issues, this method should be used to enqueue kernel signals or fault signals.

pub fn register_signalfd_poller(&self, poller: &mut PollHandle, mask: IoEvents)

pub fn prof_clock(&self) -> &Arc<ProfClock>

Returns a reference to the profiling clock of the current thread.

pub fn create_prof_timer<F>(&self, func: F) -> Arc<Timer>
where F: Fn(TimerGuard<'_>) + Send + Sync + 'static,

Creates a timer based on the profiling CPU clock of the current thread.

pub fn create_virtual_timer<F>(&self, func: F) -> Arc<Timer>
where F: Fn(TimerGuard<'_>) + Send + Sync + 'static,

Creates a timer based on the user CPU clock of the current thread.

pub fn process_expired_timers(&self)

Checks the TimerCallbacks that are managed by the prof_timer_manager. If any have timed out, call the corresponding callback functions.

pub fn credentials(&self) -> Credentials<ReadOp>

Gets the read-only credentials of the thread.

pub fn credentials_dup(&self) -> Credentials<ReadDupOp>

Gets the duplicatable read-only credentials of the thread.

pub fn credentials_mut(&self) -> Credentials<WriteOp>

Gets the write-only credentials of the current thread.

It is illegal to mutate the credentials from a thread other than the current thread. For performance reasons, this function only checks it using debug assertions.

pub fn io_priority(&self) -> &AtomicU32

Returns the I/O priority value of the thread.

pub fn ns_proxy(&self) -> &Mutex<Option<Arc<NsProxy>>>

Returns the namespaces which the thread belongs to.

pub fn timer_slack_ns(&self) -> u64

Returns the current timer slack value in nanoseconds.

pub fn set_timer_slack_ns(&self, slack_ns: u64)

Sets the current timer slack value in nanoseconds.

pub fn reset_timer_slack_to_default(&self)

Resets the current timer slack to the default value.