aster_kernel::process::posix_thread

Module thread_local

Structs

ThreadFpu

The FPU information for the current thread.

ThreadLocal

Local data for a POSIX thread.

ThreadLocalOptionRef

An immutable, shared reference to thread-local data contained within a RefCell<Option<..>>.

ThreadLocalOptionRefMut

A mutable, exclusive reference to thread-local data contained within a RefCell<Option<..>>.

Enums

FpuState 🔒

The current state of ThreadFpu.

Traits

AsThreadLocal

A trait to provide the as_thread_local method for tasks.

Type Aliases

FileTableRef

An immutable, shared reference to the file table in ThreadLocal.

FileTableRefMut

A mutable, exclusive reference to the file table in ThreadLocal.

NsProxyRef

An immutable, shared reference to the NsProxy in ThreadLocal.

NsProxyRefMut 🔒

A mutable, exclusive reference to the NsProxy in ThreadLocal.