aster_kernel::process::posix_thread

Module thread_table

Statics

THREAD_TABLE 🔒

Functions

add_thread

Adds a posix thread to global thread table

get_thread

Gets a posix thread from the global thread table

make_current_main_thread 🔒

Makes the current thread become the main thread if necessary.

remove_thread

Removes a posix thread to global thread table