aster_kernel::process

Module process

Re-exports

pub use init_proc::spawn_init_process;

pub use job_control::JobControl;

pub use process_group::ProcessGroup;

pub use session::Session;

pub use terminal::Terminal;

Modules

init_proc 🔒

This module defines functions related to spawning the init process.

job_control 🔒

process_group 🔒

session 🔒

terminal 🔒

timer_manager 🔒

Structs

AtomicPid

ParentProcess

Representing a parent process by holding a weak reference to it and its PID.

ParentProcessGuard

Process

Process stands for a set of threads that shares the same userspace.

ReapedChildrenStats

The resource usage statistics of child processes that have terminated and been reaped by the parent.

Functions

broadcast_signal_async

Broadcasts a process-directed kernel signal asynchronously.

enqueue_signal_async

Enqueues a process-directed kernel signal asynchronously.

init_in_first_process 🔒

init_on_each_cpu 🔒

Type Aliases

ExitCode

Pgid

Process group id.

Pid

Process id.

Sid

Session Id.