aster_kernel::process::process

Struct Process

pub struct Process {

    pid: u32,
    vmar: Mutex<Option<Arc<Vmar>>>,
    children_wait_queue: WaitQueue,
    pub(super) pidfile_pollee: Pollee,
    tasks: Mutex<TaskSet>,
    status: ProcessStatus,
    pub(super) parent: ParentProcess,
    children: Mutex<Option<BTreeMap<u32, Arc<Process>>>>,
    pub(super) process_group: Mutex<Weak<ProcessGroup>>,
    reaped_children_stats: Mutex<ReapedChildrenStats>,
    resource_limits: ResourceLimits,
    cgroup: RcuOption<Arc<CgroupNode>>,
    nice: AtomicNice,
    oom_score_adj: AtomicI16,
    is_child_subreaper: AtomicBool,
    pub(super) has_child_subreaper: AtomicBool,
    sig_dispositions: Mutex<Arc<Mutex<SigDispositions>>>,
    sig_queues: SigQueues,
    parent_death_signal: AtomicSigNum,
    exit_signal: AtomicSigNum,
    prof_clock: Arc<ProfClock>,
    timer_manager: PosixTimerManager,
    user_ns: Mutex<Arc<UserNamespace>>,

}

Process stands for a set of threads that shares the same userspace.

Fields
pid: u32vmar: Mutex<Option<Arc<Vmar>>>children_wait_queue: WaitQueue

Wait for child status changed

pidfile_pollee: Polleetasks: Mutex<TaskSet>

The threads

status: ProcessStatus

Process status

parent: ParentProcess

Parent process

children: Mutex<Option<BTreeMap<u32, Arc<Process>>>>

Children processes

process_group: Mutex<Weak<ProcessGroup>>

Process group

reaped_children_stats: Mutex<ReapedChildrenStats>

The resource usage statistics of reaped child processes.

resource_limits: ResourceLimits

resource limits

cgroup: RcuOption<Arc<CgroupNode>>

The bound cgroup of the process.

If this field is None, the process is bound to the root cgroup.

nice: AtomicNice

Scheduling priority nice value According to POSIX.1, the nice value is a per-process attribute, the threads in a process should share a nice value.

oom_score_adj: AtomicI16

The adjustment value of the out-of-memory (OOM) killer score.

is_child_subreaper: AtomicBool

Whether the process is a child subreaper.

A subreaper can be considered as a sort of “sub-init”. Instead of letting the init process to reap all orphan zombie processes, a subreaper can reap orphan zombie processes among its descendants.

has_child_subreaper: AtomicBool

Whether the process has a subreaper that will reap it when the process becomes orphaned.

If has_child_subreaper is true in a Process, this attribute should also be true for all of its descendants.

sig_dispositions: Mutex<Arc<Mutex<SigDispositions>>>

Sig dispositions

sig_queues: SigQueues

The process-level sigqueue.

parent_death_signal: AtomicSigNum

The signal that the process should receive when parent process exits.

exit_signal: AtomicSigNum

The signal that should be sent to the parent when this process exits.

prof_clock: Arc<ProfClock>

A profiling clock measures the user CPU time and kernel CPU time of the current process.

timer_manager: PosixTimerManager

A manager that manages timer resources and utilities of the process.

user_ns: Mutex<Arc<UserNamespace>>

The user namespace

Implementations

impl Process

pub fn current() -> Option<Arc<Process>>

Returns the current process.

It returns None if:

the function is called in the bootstrap context;

or if the current task is not associated with a process.

pub(super) fn new( pid: u32, vmar: Arc<Vmar>, resource_limits: ResourceLimits, nice: Nice, oom_score_adj: i16, sig_dispositions: Arc<Mutex<SigDispositions>>, user_ns: Arc<UserNamespace>, ) -> Arc<Self>

pub(super) fn run(&self)

Runs the process.

pub fn pid(&self) -> u32

pub fn prof_clock(&self) -> &Arc<ProfClock>

Gets the profiling clock of the process.

pub fn timer_manager(&self) -> &PosixTimerManager

Gets the timer resources and utilities of the process.

pub fn tasks(&self) -> &Mutex<TaskSet>

pub fn resource_limits(&self) -> &ResourceLimits

pub fn nice(&self) -> &AtomicNice

pub fn main_thread(&self) -> Arc<Thread>

pub fn oom_score_adj(&self) -> &AtomicI16

pub fn parent(&self) -> &ParentProcess

pub fn is_init_process(&self) -> bool

pub(super) fn children(&self) -> &Mutex<Option<BTreeMap<u32, Arc<Process>>>>

pub fn children_wait_queue(&self) -> &WaitQueue

pub fn reaped_children_stats(&self) -> &Mutex<ReapedChildrenStats>

pub fn pgid(&self) -> u32

Returns the process group ID of the process.

pub fn sid(&self) -> u32

Returns the session ID of the process.

pub fn terminal(&self) -> Option<Arc<dyn Terminal>>

Returns the controlling terminal of the process, if any.

pub fn to_new_session(self: &Arc<Self>) -> Result<u32, Error>

Moves the process to the new session.

This method will create a new process group in a new session, move the process to the new session, and return the session ID (which is equal to the process ID and the process group ID).

Errors

This method will return EPERM if an existing process group has the same identifier as the process ID. This means that the process is or was a process group leader and that the process group is still alive.

pub(super) fn clear_old_group_and_session( &self, process_group_mut: &mut MutexGuard<'_, Weak<ProcessGroup>>, session_table_mut: &mut MutexGuard<'_, BTreeMap<u32, Arc<Session>>>, group_table_mut: &mut MutexGuard<'_, BTreeMap<u32, Arc<ProcessGroup>>>, )

fn set_new_session( self: &Arc<Self>, process_group_mut: &mut MutexGuard<'_, Weak<ProcessGroup>>, session_table_mut: &mut MutexGuard<'_, BTreeMap<u32, Arc<Session>>>, group_table_mut: &mut MutexGuard<'_, BTreeMap<u32, Arc<ProcessGroup>>>, ) -> u32

pub fn move_process_to_group(&self, pid: u32, pgid: u32) -> Result<(), Error>

Moves the process itself or its child process to another process group.

The process to be moved is specified with the process ID pid; self is used only for permission checking purposes (see the Errors section below), which is typically current!() when implementing system calls.

If pgid is equal to the process ID, a new process group with the given PGID will be created (if it does not already exist). Then, the process will be moved to the process group with the given PGID, if the process group exists and belongs to the same session as the given process.

Errors

This method will return ESRCH in following cases:

The process specified by pid does not exist;

The process specified by pid is neither self or a child process of self.

This method will return EPERM in following cases:

The process is not in the same session as self;

The process is a session leader, but the given PGID is not the process’s PID/PGID;

The process group already exists, but the group does not belong to the same session;

The process group does not exist, but pgid is not equal to the process ID.

fn to_existing_group( self: &Arc<Self>, current_session: Option<Arc<Session>>, group_table_mut: MutexGuard<'_, BTreeMap<u32, Arc<ProcessGroup>>>, new_process_group: Arc<ProcessGroup>, ) -> Result<(), Error>

Moves the process to an existing group.

fn to_new_group( self: &Arc<Self>, current_session: Option<Arc<Session>>, group_table_mut: MutexGuard<'_, BTreeMap<u32, Arc<ProcessGroup>>>, ) -> Result<(), Error>

Creates a new process group and moves the process to the group.

pub fn lock_vmar(&self) -> ProcessVmarGuard<'_>

pub fn sig_dispositions(&self) -> &Mutex<Arc<Mutex<SigDispositions>>>

pub(super) fn sig_queues(&self) -> &SigQueues

pub fn enqueue_signal(&self, signal: Box<dyn Signal>)

Enqueues a process-directed signal.

This method does not perform permission checks on user signals. Therefore, unless the caller can ensure that there are no permission issues, this method should be used to enqueue kernel signals or fault signals.

pub fn clear_parent_death_signal(&self)

Clears the parent death signal.

pub fn set_parent_death_signal(&self, sig_num: SigNum)

Sets the parent death signal as signum.

pub fn parent_death_signal(&self) -> Option<SigNum>

Returns the parent death signal.

The parent death signal is the signal will be sent to child processes when the process exits.

pub fn set_exit_signal(&self, sig_num: SigNum)

pub fn exit_signal(&self) -> Option<SigNum>

pub fn status(&self) -> &ProcessStatus

Returns a reference to the process status.

pub fn stop(&self, sig_num: SigNum)

Stops the process.

pub fn resume(&self)

Resumes the stopped process.

pub fn is_stopped(&self) -> bool

Returns whether the process is stopped.

pub(super) fn wait_stopped_or_continued( &self, options: WaitOptions, ) -> Option<StopWaitStatus>

Gets and clears the stop status changes for the wait syscall.

fn wake_up_parent(&self)

pub fn set_child_subreaper(&self)

Sets the child subreaper attribute of the current process.

Panics

This method may panic if the process is a zombie process.

pub fn unset_child_subreaper(&self)

Unsets the child subreaper attribute of the current process.

pub fn is_child_subreaper(&self) -> bool

Returns whether this process is a child subreaper.

fn propagate_has_child_subreaper(&self)

Sets all descendants of the current process as having child subreaper.

pub fn user_ns(&self) -> &Mutex<Arc<UserNamespace>>

pub fn cgroup(&self) -> RcuOptionReadGuard<'_, Arc<CgroupNode>>

Returns a RCU read guard to the cgroup of the process.

The returned cgroup is not a stable snapshot. It may be changed by other threads and encounter race conditions. Users can use CgroupMembership to obtain a lock to prevent the cgroup from being changed.