aster_kernel::process::process::terminal

Trait Terminal

pub trait Terminal: Device {
    // Required method
    fn job_control(&self) -> &JobControl;
}

A terminal.

We currently support two kinds of terminal, the TTY and pty. They’re associated with a JobControl to track the session and the foreground process group.

Required Methods

fn job_control(&self) -> &JobControl

Returns the job control of the terminal.

Implementations

impl dyn Terminal

pub fn job_ioctl( self: Arc<Self>, raw_ioctl: RawIoctl, via_master: bool, ) -> Result<(), Error>

pub(super) fn set_control( self: Arc<Self>, process: &Process, ) -> Result<(), Error>

Sets the terminal to be the controlling terminal of the process.

fn unset_control(self: Arc<Self>, process: &Process) -> Result<(), Error>

Unsets the terminal from the controlling terminal of the process.

fn set_foreground( self: Arc<Self>, pgid: u32, process: &Process, ) -> Result<(), Error>

Sets the foreground process group of the terminal.

fn is_control_and<F, R>( self: &Arc<Self>, process: &Process, op: F, ) -> Result<R, Error>
where F: FnOnce(&Arc<Session>, &mut SessionGuard<'_>) -> Result<R, Error>,

Runs op when the process controls the terminal.

Note that this requires that the terminal is the controlling terminal of the session, but does not require that the process is a session leader.