aster_kernel::process

Module process_table

A global table stores the pid to process mapping. This table can be used to get process with pid. TODO: progress group, thread all need similar mapping

Structs

ProcessTable

Process Table.

ProcessTableIter

An iterator over the processes of the process table.

Enums

PidEvent

Statics

PROCESS_GROUP_TABLE 🔒

PROCESS_TABLE 🔒

SESSION_TABLE 🔒

Functions

contain_process_group

Returns whether process table contains process group with pgid

get_process

Gets a process with pid

get_process_group

Gets a process group with pgid

get_session

Gets a session with sid.

group_table_mut 🔒

process_num

Returns the number of current processes.

process_table_mut

register_observer

Registers an observer which watches PidEvent.

session_table_mut 🔒

unregister_observer

Unregisters an observer which watches PidEvent.