aster_kernel::process

Module process_vm

This module defines struct ProcessVm to represent the layout of user space process virtual memory.

The ProcessVm struct contains Vmar, which stores all existing memory mappings. The Vm also contains the basic info of process level vm segments, like init stack and heap.

Re-exports

pub use self::heap::Heap;

pub use self::heap::LockedHeap;

pub use self::init_stack::INIT_STACK_SIZE;

pub use self::init_stack::InitStack;

pub use self::init_stack::InitStackReader;

pub use self::init_stack::MAX_LEN_STRING_ARG;

pub use self::init_stack::MAX_NR_STRING_ARGS;

pub use self::init_stack::aux_vec::AuxKey;

pub use self::init_stack::aux_vec::AuxVec;

Modules

heap 🔒

init_stack 🔒

The init stack for the process.

Structs

ProcessVm

The process user space virtual memory

ProcessVmarGuard

A guard to the Vmar used by a process.

Functions

activate_vmar 🔒

Activates the Vmar in the current process’s context.