aster_kernel::process::process_vm

Module init_stack

The init stack for the process.

The init stack is used to store argv, envp, and auxiliary vectors. We can read argv and envp of a process from the init stack. Usually, the lowest address of the init stack is the highest address of the user stack of the main thread.

However, the init stack will be mapped to user space and the user process can write the content of the init stack, so the content reading from the init stack may not be the same as the initial one.

Modules

aux_vec

Structs

InitStack

The initial portion of the main stack of a process.

InitStackReader

A reader to parse the content of an InitStack.

InitStackWriter 🔒

A writer to initialize the content of an InitStack.

Constants

INIT_STACK_SIZE

Set the initial stack size to 8 megabytes, following the default Linux stack size limit.

MAX_LEN_STRING_ARG

The maximum size, in bytes, of a single argument or environment string (argv / envp) accepted by execve.

MAX_NR_STRING_ARGS

The maximum number of argument or environment strings that can be supplied to the execve system call.

Functions

generate_random_for_aux_vec 🔒