aster_kernel::process::program_loader::elf

Module load_elf

ELF file parser.

Structs

ElfLoadInfo

ElfMappedInfo 🔒

The information of mapped ELF segments.

LdsoLoadInfo 🔒

Constants

PIE_BASE_ADDR 🔒

The base address for PIE (ET_DYN with INTERP) loading.

Functions

init_aux_vec 🔒

load_elf_to_vmar

Loads an ELF file to the process VMAR.

load_ldso 🔒

lookup_and_parse_ldso 🔒

map_segment_vmo 🔒

Creates and maps the segment VMO to the VMAR.

map_segment_vmos 🔒

Initializes a Vmo for each segment and then map to the Vmar.

map_vdso_to_vmar 🔒

Maps the vDSO VMO to the corresponding virtual memory address.

map_vmos_and_build_aux_vec 🔒

Maps the VMOs to the corresponding virtual memory addresses and builds the auxiliary vector.