aster_kernel::process::signal::pause

Trait Pause

pub trait Pause: WaitTimeout {
    // Required method
    fn pause_timeout(&self, timeout: &TimeoutExt<'_>) -> Result<(), Error>;

    // Provided methods
    fn pause_until<F, R>(&self, cond: F) -> Result<R, Error>
       where F: FnMut() -> Option<R> { ... }
    fn pause_until_by<F, R>(
        &self,
        cond: F,
        reason: PauseReason,
    ) -> Result<R, Error>
       where F: FnMut() -> Option<R> { ... }
    fn pause_until_or_timeout<'a, F, T, R>(
        &self,
        cond: F,
        timeout: T,
    ) -> Result<R, Error>
       where F: FnMut() -> Option<R>,
             T: Into<TimeoutExt<'a>> { ... }
}

Pause is an extension trait to make Waiter and WaitQueue signal aware.

The original methods of Waiter and WaitQueue only allow a thread to wait (via wait) until it is woken up (via wake_up, wake_one, or wake_all) or a condition is met (via wait_until). The WaitTimeout extension trait grants the extra ability to wait until a timeout (via wait_until_or_timeout). On top of WaitTimeout, this Pause trait provides the pause-family methods, which are similar to the wait-family methods except that the methods also return when the waiting thread is interrupted by a POSIX signal. When this happens, the pause-family methods return Err(EINTR).

Required Methods

fn pause_timeout(&self, timeout: &TimeoutExt<'_>) -> Result<(), Error>

Pauses until the thread is woken up, the timeout is reached, or a signal interrupts.

Errors

This method will return an error with

EINTR if a signal is received;

ETIME if the timeout is reached.

Provided Methods

fn pause_until<F, R>(&self, cond: F) -> Result<R, Error>
where F: FnMut() -> Option<R>,

Pauses until the condition is met or a signal interrupts.

Errors

This method will return an error with EINTR if a signal is received before the condition is met.

fn pause_until_by<F, R>(&self, cond: F, reason: PauseReason) -> Result<R, Error>
where F: FnMut() -> Option<R>,

Pauses until the condition is met or a signal interrupts.

This pause happens due to the reason specified. If the reason is PauseReason::Sleep, the caller should use pause_until straightforwardly.

Errors

This method will return an error with EINTR if a signal is received before the condition is met.

fn pause_until_or_timeout<'a, F, T, R>( &self, cond: F, timeout: T, ) -> Result<R, Error>
where F: FnMut() -> Option<R>, T: Into<TimeoutExt<'a>>,

Pauses until the condition is met, the timeout is reached, or a signal interrupts.

Errors

Before the condition is met, this method will return an error with

EINTR if a signal is received;

ETIME if the timeout is reached.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.