aster_kernel::process::signal::pending

Trait HandlePendingSignal

pub trait HandlePendingSignal {
    // Required methods
    fn pending_signals(&self) -> SigSet;
    fn has_pending(&self) -> bool;
    fn has_pending_sigkill(&self) -> bool;
    fn dequeue_signal(&self, mask: &SigSet) -> Option<Box<dyn Signal>>;
}

Trait for handling pending signals.

Required Methods

fn pending_signals(&self) -> SigSet

Returns the thread’s pending signal set.

This includes signals that are currently blocked or ignored.

fn has_pending(&self) -> bool

Returns if there are pending signals that are neither blocked nor ignored.

Note that ignored but not blocked signals may be dequeued silently.

fn has_pending_sigkill(&self) -> bool

Returns if a SIGKILL signal is pending.

fn dequeue_signal(&self, mask: &SigSet) -> Option<Box<dyn Signal>>

Dequeues the next pending signal that is not masked by mask.

Returns None if no such signal is available.