aster_kernel::process::signal::poll

Trait Pollable

pub trait Pollable {
    // Required method
    fn poll(&self, mask: IoEvents, poller: Option<&mut PollHandle>) -> IoEvents;

    // Provided method
    fn wait_events<F, R>(
        &self,
        mask: IoEvents,
        timeout: Option<&Duration>,
        try_op: F,
    ) -> Result<R, Error>
       where Self: Sized,
             F: FnMut() -> Result<R, Error> { ... }
}

The Pollable trait allows for waiting for events and performing event-based operations.

Implementors are required to provide a method, Pollable::poll, which is usually implemented by simply calling Pollable::poll on the internal Pollee. This trait provides another method, Pollable::wait_events, to allow waiting for events and performing operations according to the events.

This trait is added instead of creating a new method in Pollee because sometimes we do not have access to the internal Pollee, but there is a method that provides the same semantics as Pollable::poll and we need to perform event-based operations using that method.

Required Methods

fn poll(&self, mask: IoEvents, poller: Option<&mut PollHandle>) -> IoEvents

Returns the interesting events now and monitors their occurrence in the future if the poller is provided.

This method has the same semantics as Pollee::poll_with.

Provided Methods

fn wait_events<F, R>( &self, mask: IoEvents, timeout: Option<&Duration>, try_op: F, ) -> Result<R, Error>
where Self: Sized, F: FnMut() -> Result<R, Error>,

Waits for events and performs event-based operations.

If a call to try_op() succeeds or fails with an error code other than EAGAIN, the method will return whatever the call to try_op() returns. Otherwise, the method will wait for some interesting events specified in mask to happen and try again.

This method will fail with ETIME if the timeout is specified and the event does not occur before the timeout expires.

The user must ensure that a call to try_op() does not fail with EAGAIN when the interesting events occur. However, it is allowed to have spurious EAGAIN failures due to race options where the events are consumed by another thread.