aster_kernel::process::sync::condvar

Struct Condvar

pub struct Condvar {
    waitqueue: Arc<WaitQueue>,
    counter: SpinLock<Inner>,
}

A Condvar (Condition Variable) is a synchronization primitive that can block threads until a certain condition becomes true.

Although a Condvar can block threads, it is primarily used to achieve thread synchronization. Threads waiting on a Condvar must acquire a mutex before proceeding. This setup is commonly used with a shared mutable state to ensure safe concurrent access. A typical use involves one or more threads waiting for a condition to become true to proceed with their operations.

Usage

Pair a Condvar with a Mutex to allow threads to wait for certain conditions safely. A waiting thread will sleep and atomically release the associated mutex. Another thread can then update the shared state and notify the Condvar, allowing the waiting thread to reacquire the mutex and proceed.

Example

This example demonstrates how a Condvar can synchronize threads:

use alloc::sync::Arc;
use ostd::sync::Mutex;
use crate::{process::sync::Condvar, thread::kernel_thread::Thread};

// Initializing a shared condition between threads
let pair = Arc::new((Mutex::new(false), Condvar::new()));
let pair2 = Arc::clone(&pair);

// Spawning a new kernel thread to change a shared state and notify the Condvar
ThreadOptions::new(move || {
    let (lock, cvar) = &*pair2;
    Thread::yield_now();
    let mut started = lock.lock();
    *started = true; // Modifying the shared state
    cvar.notify_one(); // Notifying one waiting thread
})
.spawn();

// Main thread waiting for the shared state to be set to true
{
    let (lock, cvar) = &*pair;
    let mut started = lock.lock();
    while !*started {
        started = cvar.wait(started).unwrap_or_else(|err| err.into_guard());
    }
}

In this example, the main thread and a child thread synchronize access to a boolean flag using a Mutex and a Condvar. The main thread waits for the flag to be set to true, utilizing the Condvar to sleep efficiently until the condition is met.

Fields
waitqueue: Arc<WaitQueue>counter: SpinLock<Inner>
Implementations

impl Condvar

pub fn new() -> Self

Creates a new condition variable.

pub fn wait<'a, T>( &self, guard: MutexGuard<'a, T>, ) -> Result<MutexGuard<'a, T>, LockErr<MutexGuard<'a, T>>>

Atomically releases the given MutexGuard, blocking the current thread until the condition variable is notified, after which the mutex will be reacquired.

Returns a new MutexGuard if the operation is successful, or returns the provided guard within a LockErr if the waiting operation fails.

pub fn wait_timeout<'a, T>( &self, guard: MutexGuard<'a, T>, timeout: Duration, ) -> Result<(MutexGuard<'a, T>, bool), LockErr<(MutexGuard<'a, T>, bool)>>

Waits for the condition variable to be signaled or broadcasted, or a timeout to elapse. bool is true if the timeout is reached.

The function returns a tuple containing a MutexGuard and a boolean that is true if the timeout elapsed before the condition variable was notified.

pub fn wait_timeout_while<'a, T, F>( &self, guard: MutexGuard<'a, T>, timeout: Duration, condition: F, ) -> Result<(MutexGuard<'a, T>, bool), LockErr<(MutexGuard<'a, T>, bool)>>
where F: FnMut(&mut T) -> bool,

Wait for the condition to become true, or until the timeout elapses, or until the condition is explicitly woken up. bool is true if the timeout is reached.

Similar to wait_timeout, it returns a tuple containing the MutexGuard and a boolean value indicating whether the wait operation terminated due to a timeout.

pub fn wait_while<'a, T, F>( &self, guard: MutexGuard<'a, T>, condition: F, ) -> Result<MutexGuard<'a, T>, LockErr<MutexGuard<'a, T>>>
where F: FnMut(&mut T) -> bool,

Wait for the condition to become true, and until the condition is explicitly woken up or interrupted.

This function blocks until either the condition becomes false or the condition variable is explicitly notified. Returns the MutexGuard if the operation completes successfully.

pub fn notify_one(&self)

Wakes up one blocked thread waiting on this condition variable.

If there is a waiting thread, it will be unblocked and allowed to reacquire the associated mutex. If no threads are waiting, this function is a no-op.

pub fn notify_all(&self)

Wakes up all blocked threads waiting on this condition variable.

This method will unblock all waiting threads and they will be allowed to reacquire the associated mutex. If no threads are waiting, this function is a no-op.