aster_kernel::sched

Module sched_class

Completely Fair Scheduler (CFS).

Re-exports

pub use self::policy::SchedPolicy;

pub use self::real_time::RealTimePolicy;

pub use self::real_time::RealTimePriority;

Modules

fair 🔒

idle 🔒

policy 🔒

real_time 🔒

stop 🔒

time 🔒

Structs

ClassScheduler

Represents the middle layer between scheduling classes and generic scheduler traits. It consists of all the sets of run queues for CPU cores. Other global information may also be stored here.

CurrentRuntime 🔒

Stores the runtime information of the current task.

PerCpuClassRqSet 🔒

Represents the run queue for each CPU core. It stores a list of run queues for scheduling classes in its corresponding CPU core. The current task of this CPU core is also stored in this structure.

PerCpuLoadStats 🔒

Holds per-CPU load information.

SchedAttr

The scheduling attribute for a thread.

Traits

SchedClassRq 🔒

The run queue for scheduling classes (the main trait). Scheduling classes should implement this trait to function as expected.

Functions

init

init_on_each_cpu

Type Aliases

SchedEntity 🔒