aster_kernel::sched::sched_class

Trait SchedClassRq

trait SchedClassRq: Send + Debug {
    // Required methods
    fn enqueue(&mut self, task: Arc<Task>, flags: Option<EnqueueFlags>);
    fn len(&self) -> usize;
    fn pick_next(&mut self) -> Option<Arc<Task>>;
    fn update_current(
        &mut self,
        rt: &CurrentRuntime,
        attr: &SchedAttr,
        flags: UpdateFlags,
    ) -> bool;

    // Provided method
    fn is_empty(&self) -> bool { ... }
}

The run queue for scheduling classes (the main trait). Scheduling classes should implement this trait to function as expected.

Required Methods

fn enqueue(&mut self, task: Arc<Task>, flags: Option<EnqueueFlags>)

Enqueues a task into the run queue.

fn len(&self) -> usize

Returns the number of threads in the run queue.

fn pick_next(&mut self) -> Option<Arc<Task>>

Picks the next task for running.

fn update_current( &mut self, rt: &CurrentRuntime, attr: &SchedAttr, flags: UpdateFlags, ) -> bool

Update the information of the current task.

The return value of this method indicates whether there is another task in this run queue to replace the current one.

Provided Methods

fn is_empty(&self) -> bool

Checks if the run queue is empty.