aster_kernel::sched::stats::scheduler_stats

Trait SchedulerStats

pub trait SchedulerStats: Sync + Send {
    // Required method
    fn nr_queued_and_running(&self) -> (u32, u32);
}

The trait for the scheduler statistics.

Required Methods

fn nr_queued_and_running(&self) -> (u32, u32)

Returns a tuple with the number of tasks in the runqueues and the number of running tasks.

We decided to return a tuple instead of having two separate functions to avoid the overhead of disabling the preemption twice to inspect the scheduler.