aster_kernel::syscall

Module fcntl

Structs

c_flock

C struct for a file range lock in Libc

Enums

FcntlCmd ğŸ”’

RangeLockWhence

Functions

from_c_flock_and_file ğŸ”’

Create the file range through C flock and opened file reference

handle_addseal ğŸ”’

handle_dupfd ğŸ”’

handle_getfd ğŸ”’

handle_getfl ğŸ”’

handle_getlk ğŸ”’

handle_getown ğŸ”’

handle_getseal ğŸ”’

handle_setfd ğŸ”’

handle_setfl ğŸ”’

handle_setlk ğŸ”’

handle_setown ğŸ”’

sys_fcntl

Type Aliases

off_t