aster_kernel::syscall

Module utimens

Structs

TimeSpecPair 🔒

Utimbuf 🔒

This struct is corresponding to the utimbuf struct in Linux.

UtimensFlags 🔒

Constants

UTIME_NOW 🔒

UTIME_OMIT 🔒

Traits

UtimeExt 🔒

Functions

do_futimesat 🔒

do_utimes 🔒

read_time_from_user 🔒

sys_futimesat

The ‘sys_futimesat’ system call sets the access and modification times of a file. Unlike ‘sys_utimensat’, it receives time values in the form of timeval structures, and it does not support the ‘flags’ argument.

sys_utime

The ‘sys_utime’ system call is similar to ‘sys_utimes’ but uses the older ‘utimbuf’ structure to specify times.

sys_utimensat

The ‘sys_utimensat’ system call sets the access and modification times of a file. The times are defined by an array of two timespec structures, where times[0] represents the access time, and times[1] represents the modification time. The flags argument is a bit mask that can include the following values:

sys_utimes

The ‘sys_utimes’ system call sets the access and modification times of a file. It receives time values in the form of timeval structures like ‘sys_futimesat’, but it uses the current working directory as the base directory.

vfs_utimes 🔒