aster_kernel

Module time

Re-exports

pub use core::Clock;

pub use core::timer;

pub use system_time::START_TIME;

pub use system_time::SystemTime;

pub use timer::Timer;

pub use timer::TimerManager;

Modules

clocks

core 🔒

cpu_time_stats

softirq 🔒

system_time 🔒

timerfd

wait

Structs

UnixTime

Unix time measures time by the number of seconds that have elapsed since the Unix epoch, without adjustments made due to leap seconds.

itimerspec_t

This struct is corresponding to the itimerspec struct in Linux.

itimerval_t

This struct is corresponding to the itimerval struct in Linux.

timespec_t

timeval_t

Constants

NSEC_PER_SEC 🔒

NSEC_PER_USEC 🔒

TIMER_ABSTIME

The various flags for setting POSIX.1b interval timers:

USEC_PER_SEC 🔒

Functions

init 🔒

init_on_each_cpu 🔒

Type Aliases

clockid_t

suseconds_t

time_t