aster_kernel::time::core

Trait Clock

pub trait Clock: Send + Sync {
    // Required method
    fn read_time(&self) -> Duration;
}

A trait that can abstract clocks which have the ability to read time, and has a fixed resolution.

Required Methods

fn read_time(&self) -> Duration

Read the current time of this clock.