aster_kernel::time::wait

Trait WaitTimeout

pub trait WaitTimeout {
    // Provided method
    fn wait_until_or_timeout<'a, F, T, R>(
        &self,
        cond: F,
        timeout: T,
    ) -> Result<R, Error>
       where F: FnMut() -> Option<R>,
             T: Into<TimeoutExt<'a>> { ... }
}

A trait that provide the timeout related function for Waiter and WaitQueue`.

Provided Methods

fn wait_until_or_timeout<'a, F, T, R>( &self, cond: F, timeout: T, ) -> Result<R, Error>
where F: FnMut() -> Option<R>, T: Into<TimeoutExt<'a>>,

Waits until some condition returns Some(_) or a given timeout is reached.

Errors

If the condition does not become Some(_) before the timeout is reached, this function will return an error with ETIME.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.