aster_kernel::util::copy_compact

Trait CopyCompat

pub trait CopyCompat {
    // Required methods
    fn read_val_compat<T: Pod>(
        &self,
        src: Vaddr,
        size: usize,
    ) -> Result<T, Error>;
    fn write_val_compat<'a, T: Pod>(
        &self,
        dest: Vaddr,
        size: usize,
        val: &'a T,
    ) -> Result<TrailingBytes<'a>, Error>;
}

A trait that can copies structures from/to the user space in a backward-compatible way.

Required Methods

fn read_val_compat<T: Pod>(&self, src: Vaddr, size: usize) -> Result<T, Error>

Reads a POD structure from the user space in a backward-compatible way.

The method’s behavior is described below to ensure backward compatibility:

If the size specified by the user is greater than the kernel structure’s size (i.e., size_of::<T>()), the kernel structure’s trailing part will be filled with zero.

If the size specified by the user is smaller than the kernel structure’s size (i.e., size_of::<T>()), this method will fail with Errno::E2BIG if the user structure’s trailing part contains non-zero values.

fn write_val_compat<'a, T: Pod>( &self, dest: Vaddr, size: usize, val: &'a T, ) -> Result<TrailingBytes<'a>, Error>

Writes a POD structure to the user space in a backward-compatible way.

The method’s behavior is described below to ensure backward compatibility:

If the size specified by the user is greater than the kernel structure’s size (i.e., size_of::<T>()), the user structure’s trailing part will be filled with zero.

If the size specified by the user is smaller than the kernel structure’s size (i.e., size_of::<T>()), this method will return a TrailingBytes instance and the caller can use it to check if the kernel structure’s trailing part contains non-zero values.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.