aster_kernel::util

Module ioctl

Decoding and dispatching ioctl commands.

When the system call layer handles an ioctl system call, it creates a RawIoctl, which is basically the numeric ioctl command and argument.

The component that handles the specific ioctl logic should first convert the RawIoctl to a strongly typed Ioctl instance, which decodes the ioctl command and includes type information about the ioctl argument.

To achieve this, define type aliases for Ioctl using the ioc macro, preferably in a separate module for clarity. At the function that dispatches the ioctl commands, we can glob-import the defined Ioctl type aliases and use the dispatch_ioctl macro for the ioctl dispatching.

Here is an complete example that demonstrates the basic usage:

mod ioctl_defs {
    use crate::util::ioctl::{ioc, InData, OutData, PassByVal};

    // Here we give the code in Linux to provide an intuitive guide on how to use the `ioc`
    // macro and how it corresponds to the Linux definitions. This is for demonstration
    // purposes only. For regular code, keeping the reference link below is sufficient.
    //
    // Also, note that the line containing the `ioc` macro will not be automatically formatted.
    // This is a deliberate design so that whitespaces can be manually inserted to vertically
    // align the `ioc` macro's parameters and improve code readability.

    // Reference: <https://elixir.bootlin.com/linux/v6.18/source/include/uapi/asm-generic/ioctls.h>>

    // ```c
    // #define TIOCSCTTY    0x540E
    // ```
    pub(super) type SetControlTty = ioc!(TIOCSCTTY,  0x540E,     InData<i32, PassByVal>);

    // ```c
    // #define TIOCSPTLCK   _IOW('T', 0x31, int)  /* Lock/unlock Pty */
    // ```
    pub(super) type SetPtyLock    = ioc!(TIOCSPTLCK, b'T', 0x31, InData<i32>);

    // ```c
    // #define TIOCGPTLCK   _IOR('T', 0x39, int) /* Get Pty lock state */
    //  ```
    pub(super) type GetPtyLock    = ioc!(TIOCGPTLCK, b'T', 0x39, OutData<i32>);
}

#[derive(Debug, Default)]
struct TtyState {
    is_controlling: bool,
    is_locked: bool,
}

impl TtyState {
    fn ioctl(&mut self, raw_ioctl: RawIoctl) -> Result<()> {
        use ioctl_defs::*;

        dispatch_ioctl!(match raw_ioctl {
            cmd @ SetControlTty => {
                let _should_steal_tty = cmd.get() == 1;
                self.is_controlling = true;
            }
            cmd @ SetPtyLock => {
                self.is_locked = cmd.read()? != 0;
            }
            cmd @ GetPtyLock => {
                cmd.write(if self.is_locked { &1 } else { &0 })?;
            }
            _ => return_errno_with_message!(Errno::ENOTTY, "the ioctl command is unknown"),
        });
        Ok(())
    }
}

fn fake_tty_ioctl() {
    let mut state = TtyState::default();
    assert!(!state.is_controlling);

    state.ioctl(RawIoctl::new(0x540E, 0));
    assert!(state.is_controlling);
}

Additionally, we support the following advanced usage:

For InOutData, it is possible to obtain a [SafePtr] instance to the underlying data using Ioctl::with_data_ptr. This allows some structure fields to be treated as input and others as output. TDX_CMD_GET_REPORT0 is such an example.

For OutData<[u8]> (or InData<[u8]>), the argument size is encoded in the ioctl command. A VmWriter (or VmReader) can be obtained from Ioctl::with_writer (or Ioctl::with_reader), which allows for the writing (or reading) of variable-length data. EVIOCGNAME is such an example.

For a group of ioctls that serve similar purposes, the IoctlEnum type allows to represent them all with a single type (just like an enum). EVIOCGBIT is such an example.

Modules

sealed 🔒

Macros

dispatch_ioctl 🔒

Dispatches ioctl commands.

impl_get_by_val_for 🔒

ioc 🔒

Defines an ioctl type.

Structs

InData

Input-only data.

InOutData

Input and output data, passed by pointer to a single object.

Ioctl

An ioctl command and its argument in strongly typed form.

IoctlEnum

An ioctl enum that represents one of a possible set of almost identical Ioctls.

NoData

No input/output data.

OutData

Output-only data, always passed by pointer.

RawIoctl

An ioctl command and its argument in raw form.

Enums

PassByPtr

A marker that denotes the input is passed by pointer (i.e., pointed to by the ioctl argument).

PassByVal

A marker that denotes the input is passed by value (i.e., encoded in the ioctl argument).

Functions

u16_size_of 🔒