aster_kernel::util::ioctl

Macro ioc

macro_rules! ioc {
    ($linux_name:ident, $raw:literal, $data:ty) => { ... };
    ($linux_name:ident, $magic:literal, $nr:literal, $data:ty) => { ... };
}

Defines an ioctl type.

Legacy encoding

For legacy encoding, a 16-bit raw number is specified as the ioctl command.

For example,

type NoTty = ioc!(TIOCNOTTY, 0x5422, NoData);

It is equivalent to:

type NoTty = Ioctl<0x54, 0x0E, /* IS_MODERN = */ false, NoData>;

Modern encoding

For modern encoding, a magic (type number) and a command ID (sequence number) are specified. These, along with the data direction and size, compose the 32-bit ioctl command.

For example,

type SetPtyLock = ioc!(TIOCSPTLCK, b'T', 0x31, InData<i32>);

It is equivalent to:

type SetPtyLock = Ioctl<b'T', 0x31, /* IS_MODERN = */ true, InData<i32>>;