aster_kernel::util

Module iovec

Structs

IoVec 🔒

A kernel space I/O vector.

UserIoVec 🔒

A user space I/O vector.

VmReaderArray

A collection of VmReaders.

VmWriterArray

A collection of VmWriters.

Constants

MAX_IO_VECTOR_LENGTH 🔒

The maximum number of buffers in the I/O vector.

MAX_TOTAL_IOV_BYTES 🔒

The maximum bytes of all buffers in the I/O vector.

Traits

MultiRead

Trait defining the read behavior for a collection of VmReaders.

MultiWrite

Trait defining the write behavior for a collection of VmWriters.

Functions

copy_iovs_and_convert 🔒

The util function for create VmReader/VmWriters.