aster_kernel::util::iovec

Trait MultiRead

pub trait MultiRead: ReadCString {
    // Required methods
    fn read(
        &mut self,
        writer: &mut VmWriter<'_, Infallible>,
    ) -> Result<usize, (OstdError, usize)>;
    fn sum_lens(&self) -> usize;
    fn skip_some(&mut self, nbytes: usize);

    // Provided method
    fn is_empty(&self) -> bool { ... }
}

Trait defining the read behavior for a collection of VmReaders.

Required Methods

fn read( &mut self, writer: &mut VmWriter<'_, Infallible>, ) -> Result<usize, (OstdError, usize)>

Reads the exact number of bytes required to exhaust self or fill writer, accumulating total bytes read.

If the return value is Ok(n), then n should be min(self.sum_lens(), writer.avail()).

Errors

This method returns OstdError::PageFault if a page fault occurs, along with the number of bytes copied before the error occurs. When an error is returned, both self and writer are advanced by the returned byte count.

fn sum_lens(&self) -> usize

Calculates the total length of data remaining to read.

fn skip_some(&mut self, nbytes: usize)

Skips the first nbytes bytes of data, or skips to the end if the readers have insufficient bytes.

Provided Methods

fn is_empty(&self) -> bool

Checks if the data remaining to read is empty.

Implementations

impl dyn MultiRead + '_

pub fn read_val_opt<T: Pod>(&mut self) -> Result<Option<T>, Error>

Reads a T value, returning a None if the readers have insufficient bytes.