aster_kernel::util::iovec

Trait MultiWrite

pub trait MultiWrite {
    // Required methods
    fn write(
        &mut self,
        reader: &mut VmReader<'_, Infallible>,
    ) -> Result<usize, Error>;
    fn sum_lens(&self) -> usize;
    fn skip_some(&mut self, nbytes: usize);

    // Provided method
    fn is_empty(&self) -> bool { ... }
}

Trait defining the write behavior for a collection of VmWriters.

Required Methods

fn write( &mut self, reader: &mut VmReader<'_, Infallible>, ) -> Result<usize, Error>

Writes the exact number of bytes required to exhaust writer or fill self, accumulating total bytes read.

If the return value is Ok(n), then n should be min(self.sum_lens(), reader.remain()).

Errors

This method returns Errno::EFAULT if a page fault occurs. The position of self and the reader is left unspecified when this method returns error.

fn sum_lens(&self) -> usize

Calculates the length of space available to write.

fn skip_some(&mut self, nbytes: usize)

Skips the first nbytes bytes of data, or skips to the end if the writers have insufficient bytes.

Provided Methods

fn is_empty(&self) -> bool

Checks if the space available to write is empty.

Implementations

impl dyn MultiWrite + '_

pub fn write_val_trunc<T: Pod>(&mut self, val: &T) -> Result<(), Error>

Writes a T value, truncating the value if the writers have insufficient bytes.