aster_kernel::util::net::addr::family

Function read_socket_addr_from_user

pub fn read_socket_addr_from_user(
    addr: Vaddr,
    addr_len: usize,
) -> Result<SocketAddr, Error>

Reads a socket address from userspace.

This method returns Err(EINVAL) for invalid socket address lengths and Err(EAFNOSUPPORT) for unsupported address families.

These error codes may be different from the Linux ones, but the difference is tricky and hard to fix. The main reason for this is that in Linux it’s up to each protocol to decide how to intercept the bytes that represent socket addresses, but here this method is designed to parse socket addresses before diving deep into protocol-specific code.