aster_kernel::util::net::addr::family

Function write_socket_addr_to_user

pub fn write_socket_addr_to_user(
    socket_addr: &SocketAddr,
    dest: Vaddr,
    max_len_ptr: Vaddr,
) -> Result<(), Error>

Writes a socket address and its length to userspace.

Similar to write_socket_addr_with_max_len, the socket address may be truncated if the buffer is not long enough. Even if truncation occurs, the actual length of the socket address is written to userspace. See https://man7.org/linux/man-pages/man3/recvmsg.3p.html for details on this behavior.

Panics

This method will panic if the socket address cannot be validly mapped to the corresponding Linux C structures. Currently, the only possible example is that the pathname in the UNIX domain socket address is too long.

It is guaranteed that all socket addresses returned by read_socket_addr_from_user have valid representations for the corresponding C structures, so passing them to this method will not cause panic.