aster_kernel::util::net::addr::family

Function write_socket_addr_with_max_len

pub fn write_socket_addr_with_max_len(
    socket_addr: &SocketAddr,
    dest: Vaddr,
    max_len: i32,
) -> Result<i32, Error>

Writes a socket address to the user space.

If the specified maximum length for the socket address is not enough, the socket address is truncated to the specified maximum length. This method returns the actual length of the socket address, regardless of whether the truncation occurs or not.

Panics

This method will panic if the socket address cannot be validly mapped to the corresponding Linux C structures. Currently, the only possible example is that the pathname in the UNIX domain socket address is too long.

It is guaranteed that all socket addresses returned by read_socket_addr_from_user have valid representations for the corresponding C structures, so passing them to this method will not cause panic.