aster_kernel::util::net

Module options

This module introduces utilities to support Linux get/setsockopt syscalls.

These two syscalls are used to get/set options for a socket. These options can be at different socket levels and of different types. To provide a unified interface, the Socket trait accepts a dyn SocketOption as a parameter. For each socket option, We define a struct that implements the SocketOption trait in net module.

However, different socket options may have values of different types. For example, the values can be u32, C structs, or byte arrays. Furthermore, some values may have different formats in kernel space and user space. For example, for the option ReusePort, the user space may use an i32 while the kernel space may use a bool.

We introduce the RawSocketOption trait for reading/writing socket options from/to user space. It can read/write values of different types and can convert the user type to the kernel type when reading from the user space and vice versa when writing to the user space. The RawSocketOption should not be implemented for a type by hand, and we provide macros to automatically implement the trait.

Example

Suppose we want to add a new option TcpNodelay.

First, the option should be added in the net module for the TCP socket.

impl_socket_option!(TcpNodelay(bool));

Then, we need to implement the ReadFromUser and WriteFromUser traits for the bool type in the utils module. These util functions can be shared if multiple options have the value of same type.

impl ReadFromUser for bool {
    // content omitted here
}
impl WriteFromUser for bool {
    // content omitted here
}

At last, we can implement RawSocketOption for TcpNodelay so that it can be read/from user space.

impl_raw_socket_option!(TcpNodeley);

At the syscall level, the interface is unified for all options and does not need to be modified.

Modules

ip 🔒

netlink 🔒

socket 🔒

tcp 🔒

utils 🔒

Enums

CSocketOptionLevel

Socket level.

Traits

RawSocketOption

Functions

new_raw_socket_option