aster_kernel::util::net::options

Module utils

Macros

impl_read_write_for_32bit_type 🔒

This macro is used to implement ReadFromUser and WriteToUser for u32 and i32.

Structs

CLinger 🔒

Constants

TCP_CONGESTION_NAME_MAX 🔒

Traits

ReadFromUser

Create an object by reading its C counterpart from the user space.

WriteToUser

Write an object to user space by writing its C counterpart.