aster_kernel::util::net::options::utils

Trait ReadFromUser

pub trait ReadFromUser: Sized {
    // Required method
    fn read_from_user(addr: Vaddr, max_len: u32) -> Result<Self, Error>;
}

Create an object by reading its C counterpart from the user space.

Note that the format of a value in the user space may be different from that
in the kernel space. For example, the type of a boolean value in the kernel is expressed as bool, whereas that in the user space is i32.

In addition, since the user space is not trusted by the kernel, values read from the user space must be validated by the kernel.

Required Methods

fn read_from_user(addr: Vaddr, max_len: u32) -> Result<Self, Error>

Read a struct from user space by reading its C counterpart.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.