aster_kernel::util::net::options::utils

Trait WriteToUser

pub trait WriteToUser {
    // Required method
    fn write_to_user(&self, addr: Vaddr, max_len: u32) -> Result<usize, Error>;
}

Write an object to user space by writing its C counterpart.

Note that the format of a value in the user space may be different from that
in the kernel space. But the format should be consistent with ReadFromUser, i.e, if we call read_from_user and write_from_user for the same type, the read value and write value in user space should be of same type.

Required Methods

fn write_to_user(&self, addr: Vaddr, max_len: u32) -> Result<usize, Error>