aster_kernel::util

Module read_cstring

Constants

INIT_ALLOC_SIZE 🔒

The recommended size for initial allocation.

Traits

ReadCString

A trait providing the ability to read a C string from the user space.

Functions

has_zero 🔒

Determines whether the value contains a zero byte.

is_addr_aligned 🔒

Checks if the given address is aligned.

read_until_nul_byte 🔒

Reads bytes from reader into buffer until a nul byte is found.