aster_kernel::util::read_cstring

Trait ReadCString

pub trait ReadCString {
    // Required methods
    fn read_cstring_until_nul(
        &mut self,
        max_len: usize,
    ) -> Result<Option<CString>, Error>;
    fn read_cstring_until_end(
        &mut self,
        max_len: usize,
    ) -> Result<(CString, usize), Error>;
}

A trait providing the ability to read a C string from the user space.

Required Methods

fn read_cstring_until_nul( &mut self, max_len: usize, ) -> Result<Option<CString>, Error>

Reads bytes until the first nul byte and creates a C string.

This method reads up to max_len bytes. The kernel must limit max_len to prevent unbounded heap allocation (i.e. it cannot be a value specified arbitrarily by the user space). If no nul terminator is found after exhausting the reader or reading out max_len bytes, this method will return None.

fn read_cstring_until_end( &mut self, max_len: usize, ) -> Result<(CString, usize), Error>

Reads bytes until the first nul byte or the reader end, and creates a C string.

This method reads up to max_len bytes. The kernel must limit max_len to prevent unbounded heap allocation (i.e. it cannot be a value specified arbitrarily by the user space). If no nul terminator is found after exhausting the reader or reading max_len bytes, this method will construct a C string with a nul byte appended.

Depending on whether the nul byte is found in the reader, the number of bytes read may equal the length of the C string or the length of the C string plus one (i.e., the nul byte). To distinguish the two cases, this method also returns an integer representing the number of bytes read.