aster_kernel

Module vdso

Virtual Dynamic Shared Object (vDSO).

vDSO enables user space applications to execute routines that access kernel space data without the need for user mode and kernel mode switching. This is particularly useful for frequently invoked, read-only operations such as obtaining the current time, which can be efficiently and securely handled within the user space.

This module manages the vDSO mechanism through the Vdso structure, which contains a VdsoData instance with necessary time-related information, and a Virtual Memory Object (Vmo) that encapsulates both the data and the vDSO routines. The VMO is intended to be mapped into the address space of every user space process for efficient access.

Macros

vdso_data_field_offset 🔒

Structs

ArchVdsoData 🔒

Vdso 🔒

The vDSO singleton.

VdsoData 🔒

Plain-old-data vDSO data that will be mapped to userspace.

VdsoInstant 🔒

An instant used in VdsoData

VdsoVmoLayout

Enums

VdsoClockMode 🔒

Constants

CLOCK_TAI 🔒

COARSE_RES_CLOCK_IDS 🔒

DEFAULT_CLOCK_MODE 🔒

HIGH_RES_CLOCK_IDS 🔒

PREBUILT_VDSO_LIB 🔒

The binary of a prebuilt Linux vDSO library.

VDSO_BASES 🔒

VDSO_VMO_LAYOUT

Statics

START_SECS_COUNT 🔒

VDSO 🔒

Functions

init_in_first_kthread 🔒

init_start_secs_count 🔒

Initializes the time duration from 1970-01-01 00:00:00 to the start time.

init_vdso 🔒

Initializes the vDSO singleton.

update_vdso_coarse_res_instant 🔒

Updates instants with respect to coarse-resolution clocks in vDSO data.

update_vdso_high_res_instant 🔒

Updates instants with respect to high-resolution clocks in vDSO data.

vdso_vmo 🔒

Returns the vDSO VMO.