aster_kernel

Module vm

Virtual memory (VM).

There are two primary VM abstractions:

The VMAR (used to be Virtual Memory Address Region, now an orphan initialism) represents the entire virtual address space of a process;

The VMO (Virtual Memory Object) is a set of logically contiguous memory frames that can be mapped into one virtual address range. Frames in a VMO can be non-contiguous in physical memory.

Modules

perms

vmar

User address space management.

vmo

Virtual Memory Objects (VMOs).

Statics

FRAME_ALLOCATOR 🔒

HEAP_ALLOCATOR 🔒

__GLOBAL_FRAME_ALLOCATOR_REF 🔒

__GLOBAL_HEAP_ALLOCATOR_REF 🔒

Functions

__global_heap_slot_info_from_layout 🔒

SAFETY: The name does not collide with other symbols.

mem_total

Total physical memory in the entire system in bytes.

slot_type_from_layout 🔒