aster_kernel::vm::vmar

Module interval_set

Intervals and interval sets used in VMARs.

Structs

IntervalDrain

A draining iterator that iterates over intervals in an interval set.

IntervalIter

An iterator that iterates over intervals in an interval set.

IntervalSet

A collection that contains non-overlapping intervals as items.

Traits

Interval

The interval of an item in an interval set.