aster_kernel::vm::vmar::vmar_impls::fork

Function cow_copy_pt

fn cow_copy_pt(
    src: &mut CursorMut<'_>,
    dst: &mut CursorMut<'_>,
    size: usize,
) -> usize

Sets mappings in the source page table as read-only to trigger COW, and copies the mappings to the destination page table.

The copied range starts from src’s current position with the given size. The destination range starts from dst’s current position.

The number of physical frames copied is returned.