aster_kernel::vm::vmar::vmar_impls

Struct Vmar

pub struct Vmar {
    inner: RwMutex<VmarInner>,
    vm_space: Arc<VmSpace>,
    rss_counters: [PerCpuCounter; 2],
    process_vm: ProcessVm,
}

The VMAR (used to be Virtual Memory Address Region, but now an orphan initialism).

A VMAR is the address space of a process.

Fields
inner: RwMutex<VmarInner>

VMAR inner

vm_space: Arc<VmSpace>

The attached VmSpace

rss_counters: [PerCpuCounter; 2]

The RSS counters.

process_vm: ProcessVm

The process VM

Implementations

impl Vmar

pub fn read_remote( &self, vaddr: Vaddr, writer: &mut VmWriter<'_>, ) -> Result<usize, (Error, usize)>

Reads memory from the process user space.

This method reads until one of the conditions is met:

The writer has no available space.

Reading from the process user space or writing to the writer encounters some error.

On success, the number of bytes read is returned; On error, both the error and the number of bytes read so far are returned.

The VmSpace of the process is not required to be activated on the current CPU.

pub fn write_remote( &self, vaddr: Vaddr, reader: &mut VmReader<'_>, ) -> Result<usize, (Error, usize)>

Writes memory to the process user space.

This method writes until one of the conditions is met:

The reader has no remaining data.

Reading from the reader or writing to the process user space encounters some error.

On success, the number of bytes written is returned; On error, both the error and the number of bytes written so far are returned.

The VmSpace of the process is not required to be activated on the current CPU.

pub fn fill_zeros_remote( &self, vaddr: Vaddr, len: usize, ) -> Result<usize, (Error, usize)>

Writes zeros to the process user space.

This method writes at most len bytes of zeros to the process user space. On success, the number of bytes written is returned; on error, both the error and the number of bytes written so far are returned.

The VmSpace of the process is not required to be activated on the current CPU.

fn access_remote<F>( &self, vaddr: Vaddr, len: usize, required_page_flags: PageFlags, op: F, ) -> Result<usize, (Error, usize)>
where F: FnMut(UFrame, usize) -> Result<usize, (Error, usize)>,

Accesses memory at vaddr..vaddr+len within the process user space using op.

The VmSpace of the process is not required to be activated on the current CPU. If any page in the range is not mapped or does not have the required page flags, a page fault will be handled to try to make the page accessible.

fn query_page_with_required_flags( &self, vaddr: Vaddr, required_page_flags: PageFlags, ) -> Result<UFrame, Error>

fn query_page(&self, vaddr: Vaddr) -> Result<Option<VmQueriedItem>, Error>

impl Vmar

pub fn fork_from(vmar: &Self) -> Result<Arc<Self>, Error>

Creates a new VMAR whose content is inherited from another using copy-on-write (COW) technique.

impl Vmar

pub fn new_map( &self, size: usize, perms: VmPerms, ) -> Result<VmarMapOptions<'_>, Error>

Creates a mapping into the VMAR through a set of VMAR mapping options.

Examples

use ostd::mm::PAGE_SIZE;

use crate::vm::{perms::VmPerms, vmar::Vmar, vmo::VmoOptions};

let vmar = Vmar::new();
let vmo = VmoOptions::new(10 * PAGE_SIZE).alloc().unwrap();
let target_vaddr = 0x1234000;
let real_vaddr = vmar
    // Create a 4 * PAGE_SIZE bytes, read-only mapping
    .new_map(PAGE_SIZE * 4, VmPerms::READ).unwrap()
    // Provide an optional offset for the mapping inside the VMAR
    .offset(target_vaddr)
    // Specify an optional binding VMO.
    .vmo(vmo)
    // Provide an optional offset to indicate the corresponding offset
    // in the VMO for the mapping
    .vmo_offset(2 * PAGE_SIZE)
    .build()
    .unwrap();
assert!(real_vaddr == target_vaddr);

For more details on the available options, see VmarMapOptions.

impl Vmar

pub fn handle_page_fault( &self, page_fault_info: &PageFaultInfo, ) -> Result<(), Error>

impl Vmar

pub fn protect(&self, perms: VmPerms, range: Range<usize>) -> Result<(), Error>

Change the permissions of the memory mappings in the specified range.

The range’s start and end addresses must be page-aligned.

If the range contains unmapped pages, an ENOMEM error will be returned. Note that pages before the unmapped hole are still protected.

impl Vmar

pub fn query(&self, range: Range<usize>) -> VmarQueryGuard<'_>

Finds all the mapped regions that intersect with the specified range.

impl Vmar

pub fn resize_mapping( &self, map_addr: Vaddr, old_size: usize, new_size: usize, check_single_mapping: bool, ) -> Result<(), Error>

Resizes the original mapping.

The range of the mapping goes from map_addr..map_addr + old_size to map_addr..map_addr + new_size.

If the new mapping size is smaller than the original mapping size, the extra part will be unmapped. If the new mapping is larger than the old mapping and the extra part overlaps with existing mapping, resizing will fail and return Err.

When check_single_mapping is true, this method will check whether the range of the original mapping is covered by a single VmMapping. If not, this method will return an Err.

When check_single_mapping is false, The range of the original mapping does not have to solely map to a whole VmMapping, but it must ensure that all existing ranges have a mapping. Otherwise, this method will return an Err.

pub fn remap( &self, old_addr: Vaddr, old_size: usize, new_addr: Option<Vaddr>, new_size: usize, ) -> Result<Vaddr, Error>

Remaps the original mapping to a new address and/or size.

If the new mapping size is smaller than the original mapping size, the extra part will be unmapped.

If new_addr is Some(new_addr), this method attempts to move the mapping from old_addr..old_addr + old_size to new_addr..new_addr + new_size. If any existing mappings lie within the target range, they will be unmapped before the move.

If new_addr is None, a new range of size new_size will be allocated, and the original mapping will be moved there.

Panics

This method panics if new_addr is None and new_size <= old_size. Use resize_mapping instead in this case.

impl Vmar

pub fn clear(&self)

Clears all mappings.

After being cleared, this vmar will become an empty vmar

pub fn remove_mapping(&self, range: Range<usize>) -> Result<(), Error>

Destroys all mappings that fall within the specified range in bytes.

The range’s start and end addresses must be page-aligned.

Mappings may fall partially within the range; only the overlapped portions of the mappings are unmapped.

pub fn discard_pages(&self, range: Range<usize>) -> Result<(), Error>

Discards all pages in the mappings that fall within the specified range in bytes.

The pages will be reloaded from a clean state by the page fault handler if they are accessed later.

The range’s start and end addresses must be page-aligned.

Mappings may fall partially within the range; only the pages in the overlapped portions of the mappings are discarded.

If the range contains unmapped pages, an ENOMEM error will be returned. Note that all other pages are still discarded.

impl Vmar

pub fn new(process_vm: ProcessVm) -> Arc<Self>

Creates a new VMAR.

pub fn get_rss_counter(&self, rss_type: RssType) -> usize

Returns the current RSS count for the given RSS type.

pub fn get_mappings_total_size(&self) -> usize

Returns the total size of the mappings in bytes.

pub fn vm_space(&self) -> &Arc<VmSpace>

Returns the attached VmSpace.

pub fn process_vm(&self) -> &ProcessVm

Returns the attached ProcessVm.

fn add_rss_counter(&self, rss_type: RssType, val: isize)