aster_kernel::vm

Module vmo

Virtual Memory Objects (VMOs).

Re-exports

pub use options::VmoOptions;

pub use pager::Pager;

Modules

options 🔒

Options for allocating root and child VMOs.

pager 🔒

Structs

CommitFlags

Commit Flags.

Vmo

Virtual Memory Objects (VMOs) are a type of capability that represents a range of memory pages.

VmoFlags

VMO flags.

WritableMappingStatus

The status of writable mappings of a VMO, i.e., shared mappings that may include the PROT_WRITE permission.

Enums

VmoCommitError

The error type used for commit operations of Vmo.

Functions

get_page_idx_range

Gets the page index range that contains the offset range of VMO.