aster_kernel::device::tty::driver

Trait TtyDriver

pub trait TtyDriver:
    Send
    + Sync
    + 'static {
    const DEVICE_MAJOR_ID: u32;

    // Required methods
    fn devtmpfs_path(&self, index: u32) -> Option<String>;
    fn open(tty: Arc<Tty<Self>>) -> Result<Box<dyn FileIo>, Error>
       where Self: Sized;
    fn push_output(&self, chs: &[u8]) -> Result<usize, Error>;
    fn echo_callback(&self) -> impl FnMut(&[u8]) + '_;
    fn can_push(&self) -> bool;
    fn notify_input(&self);
    fn console(&self) -> Option<&dyn AnyConsoleDevice>;
    fn on_termios_change(&self, old_termios: &CTermios, new_termios: &CTermios);
}

A TTY driver.

A driver exposes some device-specific behavior to Tty. For example, a device provides methods to write to the output buffer (see Self::push_output), where the output buffer can be the monitor if the underlying device is framebuffer, or just a ring buffer if the underlying device is pseduoterminal).

Required Associated Constants

const DEVICE_MAJOR_ID: u32

The device major ID.

Required Methods

fn devtmpfs_path(&self, index: u32) -> Option<String>

Returns the path where the TTY should appear under devtmpfs, if any.

fn open(tty: Arc<Tty<Self>>) -> Result<Box<dyn FileIo>, Error>
where Self: Sized,

Opens the TTY.

This function will be called when opening /dev/tty.

fn push_output(&self, chs: &[u8]) -> Result<usize, Error>

Pushes characters into the output buffer.

This method returns the number of bytes pushed or fails with an error if no bytes can be pushed.

fn echo_callback(&self) -> impl FnMut(&[u8]) + '_

Returns a callback function that echoes input characters to the output buffer.

Note that the implementation may choose to hold a lock during the life of the callback. During this time, calls to other methods such as Self::push_output may cause deadlocks.

fn can_push(&self) -> bool

Returns whether new characters can be pushed into the output buffer.

This method should return false if the output buffer is full.

fn notify_input(&self)

Notifies that the input buffer now has room for new characters.

This method should be called when the state of Tty::can_push changes from false to true.

fn console(&self) -> Option<&dyn AnyConsoleDevice>

Returns the console device associated with the TTY.

If the TTY is not associated with any console device, this method will return None.

fn on_termios_change(&self, old_termios: &CTermios, new_termios: &CTermios)

Notifies that the TTY termios is changed.

This method will be called with a spin lock held, so it cannot break atomic mode.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.