aster_kernel::fs::utils::systree_inode

Trait SysTreeInodeTy

pub(in fs) trait SysTreeInodeTy:
    Send
    + Sync
    + 'static {

    // Required methods
    fn new_arc(
        node_kind: SysTreeNodeKind,
        metadata: Metadata,
        mode: InodeMode,
        parent: Weak<Self>,
    ) -> Arc<Self>
       where Self: Sized + 'static;
    fn node_kind(&self) -> &SysTreeNodeKind;
    fn metadata(&self) -> &Metadata;
    fn mode(&self) -> Result<InodeMode, Error>;
    fn set_mode(&self, mode: InodeMode) -> Result<(), Error>;
    fn extension(&self) -> &Extension;
    fn parent(&self) -> &Weak<Self>;
    fn this(&self) -> Arc<Self>
       where Self: Sized + 'static;

    // Provided methods
    fn new_root(root_node: Arc<dyn SysBranchNode>) -> Arc<Self>
       where Self: Sized + 'static { ... }
    fn new_metadata(ino: u64, type_: InodeType) -> Metadata { ... }
    fn new_attr(
        attr: SysAttr,
        node: Arc<dyn SysNode>,
        parent: Weak<Self>,
    ) -> Arc<Self>
       where Self: Sized + 'static { ... }
    fn new_symlink(
        symlink: Arc<dyn SysSymlink>,
        parent: Weak<Self>,
    ) -> Arc<Self>
       where Self: Sized + 'static { ... }
    fn new_branch_dir(
        node_kind: SysTreeNodeKind,
        mode: Option<InodeMode>,
        parent: Weak<Self>,
    ) -> Arc<Self>
       where Self: Sized + 'static { ... }
    fn new_leaf_dir(
        node_kind: SysTreeNodeKind,
        mode: Option<InodeMode>,
        parent: Weak<Self>,
    ) -> Arc<Self>
       where Self: Sized + 'static { ... }
    fn lookup_node_or_attr(
        &self,
        name: &str,
        sysnode: &dyn SysBranchNode,
    ) -> Result<Arc<dyn Inode>, Error>
       where Self: Sized + 'static { ... }
    fn lookup_attr(
        &self,
        name: &str,
        sysnode: &dyn SysNode,
    ) -> Result<Arc<dyn Inode>, Error>
       where Self: Sized + 'static { ... }
    fn new_dentry_iter(&self, min_ino: u64) -> impl Iterator<Item = Dentry> + '_
       where Self: Sized + 'static { ... }
    fn is_dir(&self) -> bool { ... }

}

A trait that abstracts an inode type backed by a SysTree node, e.g., a SysFs inode and a CgroupFs inode.

The struct implementing this trait will have a default implementation for the Inode trait. Users only need to additionally implement the Inode::fs method.

Required Methods

fn new_arc( node_kind: SysTreeNodeKind, metadata: Metadata, mode: InodeMode, parent: Weak<Self>, ) -> Arc<Self>
where Self: Sized + 'static,

fn node_kind(&self) -> &SysTreeNodeKind

fn metadata(&self) -> &Metadata

fn mode(&self) -> Result<InodeMode, Error>

fn set_mode(&self, mode: InodeMode) -> Result<(), Error>

fn extension(&self) -> &Extension

fn parent(&self) -> &Weak<Self>

fn this(&self) -> Arc<Self>
where Self: Sized + 'static,

Provided Methods

fn new_root(root_node: Arc<dyn SysBranchNode>) -> Arc<Self>
where Self: Sized + 'static,

fn new_metadata(ino: u64, type_: InodeType) -> Metadata

fn new_attr( attr: SysAttr, node: Arc<dyn SysNode>, parent: Weak<Self>, ) -> Arc<Self>
where Self: Sized + 'static,

fn new_symlink(symlink: Arc<dyn SysSymlink>, parent: Weak<Self>) -> Arc<Self>
where Self: Sized + 'static,

fn new_branch_dir( node_kind: SysTreeNodeKind, mode: Option<InodeMode>, parent: Weak<Self>, ) -> Arc<Self>
where Self: Sized + 'static,

fn new_leaf_dir( node_kind: SysTreeNodeKind, mode: Option<InodeMode>, parent: Weak<Self>, ) -> Arc<Self>
where Self: Sized + 'static,

fn lookup_node_or_attr( &self, name: &str, sysnode: &dyn SysBranchNode, ) -> Result<Arc<dyn Inode>, Error>
where Self: Sized + 'static,

fn lookup_attr( &self, name: &str, sysnode: &dyn SysNode, ) -> Result<Arc<dyn Inode>, Error>
where Self: Sized + 'static,

fn new_dentry_iter(&self, min_ino: u64) -> impl Iterator<Item = Dentry> + '_
where Self: Sized + 'static,

fn is_dir(&self) -> bool

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.