aster_kernel::fs::vfs::fs_apis::registry

Trait FsType

pub trait FsType:
    Send
    + Sync
    + 'static {
    // Required methods
    fn name(&self) -> &'static str;
    fn properties(&self) -> FsProperties;
    fn create(
        &self,
        fs_creation_ctx: &FsCreationCtx<'_>,
    ) -> Result<Arc<dyn FileSystem>, Error>;
    fn sysnode(&self) -> Option<Arc<dyn SysNode>>;
}

A type of file system.

Required Methods

fn name(&self) -> &'static str

Gets the name of this FS type such as "ext4" or "sysfs".

fn properties(&self) -> FsProperties

Gets the properties of this FS type.

fn create( &self, fs_creation_ctx: &FsCreationCtx<'_>, ) -> Result<Arc<dyn FileSystem>, Error>

Creates an instance of this FS type.

fn sysnode(&self) -> Option<Arc<dyn SysNode>>

Returns a SysTree node that represents the FS type.

If a FS type is not intended to appear under SysFs, then this method returns a None.

The same result will be returned by this method when it is called multiple times.